Authenticate and Initialize Earth Engine

Authorize rgee to manage Earth Engine resources, Google Drive, and Google Cloud Storage. The ee_initialize() via web-browser will ask to sign in to your Google account and allows you to grant permission to manage resources. This function is a wrapper around rgee::ee$Initialize().

ee_Initialize(email = NULL, drive = FALSE, gcs = FALSE, quiet = FALSE)

Arguments

email
Character (optional, e.g. data.colec.fbf@gmail.com). The email name is used as a folder inside the path rgee::ee_get_earthengine_path(). This enable a multi-user support allowing to target a specific Google identity.

drive
Logical (optional). If TRUE, the drive credential will be cached in the path rgee::ee_get_earthengine_path().

gcs
logical. If TRUE, the Google Cloud Storage credential will be cached in the path rgee::ee_get_earthengine_path().

quiet
logical. Suppress info messages.

Details

ee_Initialize(...) can also manage Google drive and Google Cloud Storage resources using the R packages googledrive and googlecloudStorageR, respectively. By default, rgee does not require them, these are only necessary for exporting and importing tasks. All user credentials are saved in the directory ~/.config/earthengine/, if a user does not specified the the email argument all user credentials will be saved in a subdirectory called ndef.

See also

Examples

library(rgee)
ee_reattach() # reattach ee as a reserved word

# Simple init
ee_Initialize()

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#> 

# Advanced init
ee_Initialize(
  email = "data.colec.fbf@gmail.com",
  drive = TRUE,
  gcs = TRUE
)

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#> 

ee_user_info()

#> Error in ee_get_earthengine_path(): rgee_sessioninfo.txt does not exist, run rgee::ee_Initialize() to fixed.