Convert Earth Engine (EE) Table in a Simple Feature (sf)

Convert Earth Engine (EE) Table in a Simple Feature (sf)

ee_as_sf(x)

Arguments

x
ee.Geometry, ee.Feature or ee.FeatureCollection object to be converted into sf.

Value

An sf class object, see Details.

Details

The type of sf object returned will depend on the input GeoJSON. Sometimes you will get back a POINTS class, and sometimes a POLYGON class, etc., depending on what the structure of the GeoJSON. The reading and writing of the CRS to/from geojson is inconsistent. You can directly set the CRS by passing a valid PROJ4 string or epsg code to the crs argument in st_read.

Examples

library(ee)

#> Error in library(ee): there is no package called ‘ee’

ee_Initialize()

#> Error in py_call_impl(callable, dots$args, dots$keywords): EEException: Server returned HTTP code: 404
#> 
#> Detailed traceback: 
#>   File "/usr/local/lib/python3.5/dist-packages/ee/data.py", line 994, in getAssetRoots
#>     return send_('/buckets', {}, 'GET')
#>   File "/usr/local/lib/python3.5/dist-packages/ee/data.py", line 1196, in send_
#>     'Server returned HTTP code: %d' % response.status)

roi <- ee$Geometry$Polygon(list(
  c(-122.27577209472656, 37.891247253777074),
  c(-122.27577209472656, 37.86875557241152),
  c(-122.24040985107422, 37.86875557241152),
  c(-122.24040985107422, 37.891247253777074)
))

#> Error in py_call_impl(callable, dots$args, dots$keywords): EEException: Server returned HTTP code: 404
#> 
#> Detailed traceback: 
#>   File "/usr/local/lib/python3.5/dist-packages/ee/geometry.py", line 399, in Polygon
#>     return Geometry(Geometry._parseArgs('Polygon', 3, all_args))
#>   File "/usr/local/lib/python3.5/dist-packages/ee/computedobject.py", line 32, in __call__
#>     return type.__call__(cls, *args, **kwargs)
#>   File "/usr/local/lib/python3.5/dist-packages/ee/geometry.py", line 62, in __init__
#>     self.initialize()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/geometry.py", line 127, in initialize
#>     apifunction.ApiFunction.importApi(cls, 'Geometry', 'Geometry')
#>   File "/usr/local/lib/python3.5/dist-packages/ee/apifunction.py", line 181, in importApi
#>     cls.initialize()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/apifunction.py", line 152, in initialize
#>     signatures = data.getAlgorithms()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/data.py", line 650, in getAlgorithms
#>     return send_('/algorithms', {}, 'GET')
#>   File "/usr/local/lib/python3.5/dist-packages/ee/data.py", line 1196, in send_
#>     'Server returned HTTP code: %d' % response.status)

blocks <- ee$FeatureCollection("TIGER/2010/Blocks")

#> Error in py_call_impl(callable, dots$args, dots$keywords): EEException: Server returned HTTP code: 404
#> 
#> Detailed traceback: 
#>   File "/usr/local/lib/python3.5/dist-packages/ee/computedobject.py", line 32, in __call__
#>     return type.__call__(cls, *args, **kwargs)
#>   File "/usr/local/lib/python3.5/dist-packages/ee/featurecollection.py", line 43, in __init__
#>     self.initialize()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/featurecollection.py", line 85, in initialize
#>     super(FeatureCollection, cls).initialize()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/collection.py", line 32, in initialize
#>     apifunction.ApiFunction.importApi(cls, 'Collection', 'Collection')
#>   File "/usr/local/lib/python3.5/dist-packages/ee/apifunction.py", line 181, in importApi
#>     cls.initialize()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/apifunction.py", line 152, in initialize
#>     signatures = data.getAlgorithms()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/data.py", line 650, in getAlgorithms
#>     return send_('/algorithms', {}, 'GET')
#>   File "/usr/local/lib/python3.5/dist-packages/ee/data.py", line 1196, in send_
#>     'Server returned HTTP code: %d' % response.status)

subset <- blocks$filterBounds(roi)

#> Error in eval(expr, envir, enclos): object 'blocks' not found

sf_subset <- ee_as_sf(subset)

#> Error in py_call_impl(callable, dots$args, dots$keywords): EEException: Server returned HTTP code: 404
#> 
#> Detailed traceback: 
#>   File "/usr/local/lib/python3.5/dist-packages/ee/computedobject.py", line 32, in __call__
#>     return type.__call__(cls, *args, **kwargs)
#>   File "/usr/local/lib/python3.5/dist-packages/ee/featurecollection.py", line 43, in __init__
#>     self.initialize()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/featurecollection.py", line 85, in initialize
#>     super(FeatureCollection, cls).initialize()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/collection.py", line 32, in initialize
#>     apifunction.ApiFunction.importApi(cls, 'Collection', 'Collection')
#>   File "/usr/local/lib/python3.5/dist-packages/ee/apifunction.py", line 181, in importApi
#>     cls.initialize()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/apifunction.py", line 152, in initialize
#>     signatures = data.getAlgorithms()
#>   File "/usr/local/lib/python3.5/dist-packages/ee/data.py", line 650, in getAlgorithms
#>     return send_('/algorithms', {}, 'GET')
#>   File "/usr/local/lib/python3.5/dist-packages/ee/data.py", line 1196, in send_
#>     'Server returned HTTP code: %d' % response.status)

plot(sf_subset)

#> Error in plot(sf_subset): object 'sf_subset' not found