Convert an Earth Engine (EE) image into a stars object

Convert an Earth Engine (EE) image into a stars object

ee_as_stars(image, region, scale = NULL, geodesic = NULL,
  evenOdd = NULL, maxPixels = 1e+09, via = "getInfo",
  container = "rgee_backup", quiet = FALSE)

Arguments

image
ee$Image to be converted into a sf object

region
EE Geometry Rectangle (ee$Geometry$Rectangle). The CRS needs to be the same that the x argument otherwise it will be forced. If not specified, image bounds will be taken.

scale
The resolution in meters per pixel. If scale is set to NULL, the native resolution of the first band will be taken.

geodesic
Whether line segments of region should be interpreted as spherical geodesics. If FALSE, indicates that line segments should be interpreted as planar lines in the specified CRS. If not specified in the geometry (region argument) defaults to TRUE if the CRS is geographic (including the default EPSG:4326), or to FALSE if the CRS is projected.

evenOdd
If TRUE, polygon interiors will be determined by the even/odd rule, where a point is inside if it crosses an odd number of edges to reach a point at infinity. Otherwise polygons use the left-inside rule, where interiors are on the left side of the shell's edges when walking the vertices in the given order. If unspecified in the geometry (region argument) defaults to TRUE.

maxPixels
The maximum allowed number of pixels in the exported image. The task will fail if the exported region covers more pixels in the specified projection. Defaults to 100,000,000.

via
Method to download the image. Three methods are implemented 'getInfo', 'drive' and 'gcs'. See details.

container
Relevant when the "via" argument is defined as 'drive' or 'gcs'. It is the name of a unique folder ('drive') or bucket ('gcs') to export into.

quiet
logical. Suppress info message

Value

A stars object, when via argument was defined as 'getInfo'. For other cases will return a stars-proxy object.

Details

The process to pass a ee$Image to your local env could be carried out by three different strategies. The first one ('getInfo') use the getInfo method, which fetch and return information about Earth Engine objects, the advantage of use this is a direct and fast download. However, there is a limit of 262144 pixels (512x512) that can be transferred by request which makes it unsatisfactory for large images. The second ('drive') and third ('gcs') method are suitable for large images since it use Google Drive and Google Cloud Storage as a intermediate container.

Examples

if (FALSE) {
library(raster)
library(stars)
library(rgee)

# Initialize a specific Earth Engine account and load
# either Google Drive or Google Cloud Storage credentials
ee_reattach()
ee_Initialize(
  email = "data.colec.fbf@gmail.com",
  drive = TRUE,
  gcs = TRUE
)
ee_user_info()

# Define an image.
img <- ee$Image("LANDSAT/LC08/C01/T1_SR/LC08_038029_20180810")$
  select(c("B4", "B3", "B2"))$
  divide(10000)

# OPTIONAL display it using Map
Map$centerObject(eeObject = img)
Map$addLayer(eeObject = img, visParams = list(max = 0.4,gamma=0.1))

# Define an area of interest.
geometry <- ee$Geometry$Rectangle(c(-110.8, 44.6, -110.6, 44.7))

## getInfo - Method 01
img_stars_01 <- ee_as_stars(
  image = img,
  region = geometry,
  geodesic = FALSE,
  via = "getInfo"
)

## drive - Method 02
img_stars_02 <- ee_as_stars(
  image = img,
  region = geometry,
  geodesic = FALSE,
  via = "drive"
)

## gcs - Method 03
img_stars_03 <- ee_as_stars(
  image = img,
  region = geometry,
  geodesic = FALSE,
  container = "rgee_dev",
  via = "gcs"
)

# OPTIONAL: Delete containers
ee_clean_container(
  name = "rgee_backup",
  type = "drive"
)
ee_clean_container(
  name = "rgee_dev",
  type = "gcs"
)
}