Move results from Google Drive to a local directory

Move results of an EE task saved in Google Drive to a local directory.

ee_drive_to_local(
  task,
  dsn,
  overwrite = TRUE,
  consider = interactive(),
  quiet = FALSE
)

Arguments

task
List generated after finished correctly a EE task. See details.

dsn
Character. Output filename. If missing, a temporary file will be assigned.

overwrite
A boolean argument which indicates indicating whether "filename" should be overwritten. By default TRUE.

consider
Interactive. See details.

quiet
logical. Suppress info message

Value

filename character vector.

Details

The task argument needs a status as task "COMPLETED" to work, since the parameters necessary to identify EE objects into google drive are obtained from ee$batch$Export$*$toDrive(...)$start()$status().

consider argument is necessary since Google Drive permits users to create files with the same name. consider uses an interactive R session by default to help users identify just the files that they want to download. Additionally, the options "last" and "all" are implemented. "last" will download just the last file saved in Google Drive while with "all" all files will be downloaded.

Examples

if (FALSE) {
library(rgee)
library(stars)
library(sf)

ee_users()
ee_reattach() # reattach ee as a reserved word
ee_Initialize(drive = TRUE)

# Define study area (local -> earth engine)
# Communal Reserve Amarakaeri - Peru
rlist <- list(xmin = -71.13, xmax = -70.95,ymin = -12.89, ymax = -12.73)
ROI <- c(rlist$xmin, rlist$ymin,
         rlist$xmax, rlist$ymin,
         rlist$xmax, rlist$ymax,
         rlist$xmin, rlist$ymax,
         rlist$xmin, rlist$ymin)

ee_ROI <- matrix(ROI, ncol = 2, byrow = TRUE) %>%
  list() %>%
  st_polygon() %>%
  st_sfc() %>%
  st_set_crs(4326) %>%
  sf_as_ee()

# Get the mean annual NDVI for 2011
cloudMaskL457 <- function(image) {
  qa <- image$select("pixel_qa")
  cloud <- qa$bitwiseAnd(32L)$
    And(qa$bitwiseAnd(128L))$
    Or(qa$bitwiseAnd(8L))
  mask2 <- image$mask()$reduce(ee$Reducer$min())
  image <- image$updateMask(cloud$Not())$updateMask(mask2)
  image$normalizedDifference(list("B4", "B3"))
}

ic_l5 <- ee$ImageCollection("LANDSAT/LT05/C01/T1_SR")$
  filterBounds(ee_ROI)$
  filterDate("2011-01-01", "2011-12-31")$
  map(cloudMaskL457)

# Create simple composite
mean_l5 <- ic_l5$mean()$rename("NDVI")
mean_l5 <- mean_l5$reproject(crs = "EPSG:4326", scale = 500)
mean_l5_Amarakaeri <- mean_l5$clip(ee_ROI)

# Move results from Earth Engine to Drive
task_img <- ee_image_to_drive(
  image = mean_l5_Amarakaeri,
  folder = "Amarakaeri",
  fileFormat = "GEO_TIFF",
  region = ee_ROI$geometry(),
  fileNamePrefix = paste0("my_image", Sys.time())
)

task_img$start()
ee_monitoring(task_img)

# Move results from Drive to local
img <- ee_drive_to_local(task = task_img)
}