Extract values from EE Images or ImageCollections objects

Extract values from a ee$Image or ImageCollection at the locations of a geometry object. You can use ee$Geometry$*, ee$Feature, ee$FeatureCollection and sf objects. This function try to mimics how extract currently works.

ee_extract(x, y, fun = ee$Reducer$mean(), scale = 1000, sf = FALSE, ...)

Arguments

x
EE Image or ImageCollection.

y
ee$Geometry$*, ee$Feature, ee$FeatureCollection or sf objects.

fun
ee$Reducer object. Function to summarize the values. The function should take a single numeric vector as argument and return a single value. See details.

scale
A nominal scale in meters of the Image projection to work in. By default 1000.

sf
Logical. Should the extracted values be added to the data.frame of the sf object y? This only applies if y is a sf object.

...
reduceRegions additional parameters. See ee_help(ee$Image()$reduceRegions) for more details.

Details

In Google Earth Engine the reducer functions that return one value are:

allNonZero: Returns a Reducer that returns 1 if all of its inputs are non-zero, 0 otherwise.

anyNonZero: Returns a Reducer that returns 1 if any of its inputs are non-zero, 0 otherwise.

bitwiseAnd: Returns a Reducer that computes the bitwise-and summation of its inputs.

bitwiseOr: Returns a Reducer that computes the bitwise-or summation of its inputs.

count: Returns a Reducer that computes the number of non-null inputs.

first: Returns a Reducer that returns the first of its inputs.

firstNonNull: Returns a Reducer that returns the first of its non-null inputs.

kurtosis: Returns a Reducer that Computes the kurtosis of its inputs.

last: Returns a Reducer that returns the last of its inputs.

lastNonNull: Returns a Reducer that returns the last of its non-null inputs.

max: Creates a reducer that outputs the maximum value of its (first) input. If numInputs is greater than one, also outputs the corresponding values of the additional inputs.

mean: Returns a Reducer that computes the (weighted) arithmetic mean of its inputs.

median: Create a reducer that will compute the median of the inputs. For small numbers of inputs (up to maxRaw) the median will be computed directly; for larger numbers of inputs the median will be derived from a histogram.

min: Creates a reducer that outputs the minimum value of its (first) input. If numInputs is greater than one, also outputs additional inputs.

mode: Create a reducer that will compute the mode of the inputs. For small numbers of inputs (up to maxRaw) the mode will be computed directly; for larger numbers of inputs the mode will be derived from a histogram.

product: Returns a Reducer that computes the product of its inputs.

sampleStdDev: Returns a Reducer that computes the sample standard deviation of its inputs.

sampleVariance: Returns a Reducer that computes the sample variance of its inputs.

stdDev: Returns a Reducer that computes the standard deviation of its inputs.

sum: Returns a Reducer that computes the (weighted) sum of its inputs.

variance: Returns a Reducer that computes the variance of its inputs.

Examples

library(rgee)
library(sf)

#> Linking to GEOS 3.7.2, GDAL 2.4.2, PROJ 5.2.0

ee_reattach() # reattach ee as a reserved word

#> Error in py_module_import(module, convert = convert): ImportError: No module named ee

ee_Initialize()

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#> 

# Define a Image or ImageCollection: Terraclimate
terraclimate <- ee$ImageCollection("IDAHO_EPSCOR/TERRACLIMATE")$
  filterDate("2001-01-01", "2002-01-01")$
  map(function(x) x$select("pr"))

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#> 

# Define a geometry
nc <- st_read(
  dsn = system.file("shape/nc.shp", package = "sf"),
  stringsAsFactors = FALSE,
  quiet = TRUE
)

# Extract values
ee_nc_rain <- ee_extract(
  x = terraclimate,
  y = nc,
  scale = 250,
  fun = ee$Reducer$max(),
  sf = TRUE
)

#> Error in py_module_import(module, convert = convert): ImportError: No module named ee
#> 
#> Detailed traceback: 
#>   File "/Users/runner/runners/2.169.1/work/_temp/Library/rgee/python/ee_extract.py", line 3, in <module>
#>     import ee

# Spatial plot
plot(ee_nc_rain["X200106"],
  main = "2001 Jan Precipitation - Terraclimate",
  reset = FALSE
)

#> Error in plot(ee_nc_rain["X200106"], main = "2001 Jan Precipitation - Terraclimate",     reset = FALSE): object 'ee_nc_rain' not found

dev.off()

#> null device 
#>           1 

# Temporal plot
ee_nc <- ee_nc_rain

#> Error in eval(expr, envir, enclos): object 'ee_nc_rain' not found

time_serie <- as.numeric(ee_nc[1, sprintf("X%s", 200101:200112)])

#> Error in eval(expr, envir, enclos): object 'ee_nc' not found

main <- sprintf("2001 Precipitation - %s", ee_nc$NAME[1])

#> Error in sprintf("2001 Precipitation - %s", ee_nc$NAME[1]): object 'ee_nc' not found

plot(time_serie, ylab = "pp (mm/month)", type = "l", lwd = 1.5, main = main)

#> Error in plot(time_serie, ylab = "pp (mm/month)", type = "l", lwd = 1.5,     main = main): object 'time_serie' not found

points(time_serie, pch = 20, lwd = 1.5, cex = 1.5)

#> Error in points(time_serie, pch = 20, lwd = 1.5, cex = 1.5): object 'time_serie' not found