Get the date of a EE ImageCollection

Get the date of a EE ImageCollection

ee_get_date_ic(x, time_end = FALSE)

Arguments

x
ee$ImageCollection object

time_end
Logical. If TRUE, the system:time_end property is also returned. See details.

Value

A data.frame with the columns: id (ID of the image), time_start, and time_end (only if the argument time_end is set as TRUE). The number of rows depends on the number of images (ee$ImageCollection$size).

Details

system:time_start Sets the start period of data acquisition while system:time_end does the same for the end period. See the Earth Engine glossary for getting more information.

See also

Examples

if (FALSE) {
library(rgee)
library(sf)
ee_Initialize()

nc <- st_read(system.file("shape/nc.shp", package = "sf")) %>%
  st_transform(4326) %>%
  sf_as_ee()

ee_s2 <- ee$ImageCollection("COPERNICUS/S2")$
  filterDate("2016-01-01", "2016-01-31")$
  filterBounds(nc)

ee_get_date_ic(ee_s2)
}