Get the date of a EE Image

Get the date of a EE Image

ee_get_date_img(x, time_end = FALSE)

Arguments

x
ee$Image or ee$ImageCollection object

time_end
Logical. If TRUE, the system:time_end property is also returned. See details.

Value

An List object with the elements: id, time_start and time_end (if the time_end argument is TRUE).

Details

system:time_start set the start period of data acquisition while system:time_end does the same for the end period. See the Earth Engine glossary for getting more information.

See also

Examples

if (FALSE) {
library(rgee)
ee_Initialize()

l8 <- ee$Image('LANDSAT/LC08/C01/T1_TOA/LC08_044034_20140318')
ee_get_date_img(l8)
srtm <- ee$Image('CGIAR/SRTM90_V4')
ee_get_date_img(srtm, time_end = TRUE)
}