Approximate size of an EE Image object

Get the approximate number of rows, cols, and size of a single-band Earth Engine Image.

ee_image_info(
  image,
  band_metadata = NULL,
  getsize = TRUE,
  compression_ratio = 20,
  quiet = FALSE
)

Arguments

image

Single-band EE Image object.

band_metadata

A list with image properties. If NULL it will be automatically generated.

getsize

Logical. If TRUE, the function will estimate the size of the object

compression_ratio

Numeric. Measurement of the relative data size reduction produced by a data compression algorithm (ignored if getsize is FALSE). By default is 20.

quiet

Logical. Suppress info message

Value

A list containing information about the number of rows (nrow), number of columns (ncol), total number of pixels (total_pixel), and image size (image_size).

Examples

if (FALSE) { # \dontrun{
library(rgee)
ee_Initialize()

# World SRTM
srtm <- ee$Image("CGIAR/SRTM90_V4")
ee_image_info(srtm)

# Landast8
l8 <- ee$Image("LANDSAT/LC08/C01/T1_SR/LC08_038029_20180810")$select("B4")
ee_image_info(l8)
} # }