Install Earth Engine API packages

Wrapping to reticulate::py_install(packages = "ee")

ee_install(packages, envname = NULL, method = c("auto", "virtualenv",
  "conda"), conda = "auto", python_version = NULL, ...)

Arguments

packages
Character vector with package names to install.

envname
The name, or full path, of the environment in which Python packages are to be installed. When NULL (the default), the active environment as set by the RETICULATE_PYTHON_ENV variable will be used; if that is unset, then the r-reticulate environment will be used.

method
Installation method. By default, "auto" automatically finds a method that will work in the local environment. Change the default to force a specific installation method. Note that the "virtualenv" method is not available on Windows.

conda
conda Path to conda executable (or "auto" to find conda using the PATH and other conventional install locations).

python_version
The requested Python version. Ignored when attempting to install with a Python virtual environment.

...
Additional arguments passed to conda_install() or virtualenv_install().

Details

On Linux and OS X the "virtualenv" method will be used by default ("conda" will be used if virtualenv isn't available). On Windows, the "conda" method is always used.