Install rgee Python packages dependencies

Install the necessary Python packages to be used in rgee.

ee_install_python_packages(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  ee_version = NULL,
  envname = NULL,
  pip = TRUE,
  conda_python_version = "3.7",
  confirm = interactive(),
  ...
)

Arguments

method
Installation method. By default, "auto" automatically finds a method that will work in the local environment. Change the default to force a specific installation method. Note that the "virtualenv" method is not available on Windows (as this isn't supported by rgee). Note also that since this command runs without privilege the "system" method is available only on Windows.

conda
Path to conda executable (or "auto" to find conda using the PATH and other conventional install locations).

ee_version
earthengine-api version to install. When this argument is NULL (the default), the version with which rgee was built will be installed.

envname
Name of Python environment, or full path, which Python packages are to be installed.

pip
Logical. Use pip for package installation? This is only relevant when Conda environments are used, as otherwise packages will be installed from the Conda repositories.

conda_python_version
the Python version installed in the created conda environment. Python 3.7 is installed by default.

confirm
Logical. Confirm if restart R when the 'install' argument is TRUE.

...
other arguments passed to conda_install or virtualenv_install.

Details

It is neccessary restart R to observe change when installing Python packages. rgee only is compatible with Python version 3.5 >=.

Examples

if (FALSE) {
library(rgee)
ee_create_pyenv('ee')
ee_discover_pyenvs()
ee_set_pyenv(
  python_path = ".../ee/bin/python",
  python_env = 'ee')
ee_install_python_packages()
}