Interface for manage the Earth Engine Asset

R functions for managing the Earth Engine Asset

ee_manage_create(path_asset, asset_type = "folder", quiet = FALSE)

ee_manage_delete(path_asset, quiet = FALSE)

ee_manage_assetlist(path_asset, quiet = FALSE)

ee_manage_quota()

ee_manage_size(path_asset)

ee_manage_copy(path_asset, final_path, quiet = FALSE)

ee_manage_move(path_asset, final_path, quiet = FALSE)

ee_manage_set_properties(path_asset, properties)

ee_manage_delete_properties(path_asset, property)

ee_manage_assets_access(path_asset,
  acl = getOption("rgee.manage.getAssetAcl"), quiet = FALSE)

ee_manage_task(quiet, cache = TRUE)

ee_manage_cancel_all_running_taks()

Arguments

path_asset
Character. Name of the EE asset e.g. users/user/nameofcollection.

asset_type
Character. The asset type ('folder' or 'imagecollection').

quiet
Logical. Suppress info message.

final_path
Character. Name of the final EE asset to copy or move.

properties
List. Set of parameters to established as a property of an EE object. See details.

property
Character. Name of a specific property to be deleted.

acl
A list describing the asset's Access Control List. See getOption("rgee.manage.getAssetAcl").

cache
Logical. TRUE save the report in the /temp directory.

Examples

library(rgee)
# Change the user and google account to be able to reproduce
ee_Initialize(user_gmail = 'csaybar@gmail.com')

#> Requesting Earth Engine authorization  ... DONE

ee_manage_create('users/csaybar/rgee')

#> GEE asset: users/csaybar/rgee created

# 1. List all the elements inside a folder or a ImageCollection
ee_manage_assetlist(path_asset = 'users/csaybar/rgee')

#> data frame with 0 columns and 0 rows

# 2. Create a Folder or a ImageCollection
ee_manage_create('users/csaybar/rgee/rgee_folder',asset_type = 'folder')

#> GEE asset: users/csaybar/rgee/rgee_folder created

ee_manage_create('users/csaybar/rgee/rgee_ic',asset_type = 'imagecollection')

#> GEE asset: users/csaybar/rgee/rgee_ic created

ee_manage_assetlist('users/csaybar/rgee')

#>                               ID            TYPE
#> 2     users/csaybar/rgee/rgee_ic ImageCollection
#> 1 users/csaybar/rgee/rgee_folder          Folder

# 3. Shows your Earth Engine quota
ee_manage_quota()

#>  Total Quota: 250.00 GB 
#>  Used Quota: 778.14 MB

# 4. Estimate the size of a Image, ImageCollection, Table or Folder.
ee_manage_size('users/csaybar/rgee')

#> Folder :
#>  - Size: 0 B
#>  - # elements: 2

# 5. Move a EE object to another folder
ee_manage_move(path_asset = 'users/csaybar/rgee/rgee_ic',
               final_path = 'users/csaybar/rgee/rgee_folder/rgee_ic_moved')

#> Done

ee_manage_assetlist('users/csaybar/rgee/rgee_folder')

#>                                             ID            TYPE
#> 1 users/csaybar/rgee/rgee_folder/rgee_ic_moved ImageCollection

# 6. Set properties to an EE object.
ee_manage_set_properties(path_asset = 'users/csaybar/rgee/rgee_folder/rgee_ic_moved',
                         properties = list(message='hello-world',language = 'R'))
test_ic <- ee$ImageCollection('users/csaybar/rgee/rgee_folder/rgee_ic_moved')
test_ic$getInfo()

#> $features
#> list()
#> 
#> $id
#> [1] "users/csaybar/rgee/rgee_folder/rgee_ic_moved"
#> 
#> $properties
#> $properties$language
#> [1] "R"
#> 
#> $properties$message
#> [1] "hello-world"
#> 
#> 
#> $version
#> [1] 1878938063
#> 
#> $type
#> [1] "ImageCollection"
#> 
#> $bands
#> list()
#> 

# 7. Delete properties
ee_manage_delete_properties(path_asset = 'users/csaybar/rgee/rgee_folder/rgee_ic_moved',
                            property = c("message","language"))
test_ic$getInfo()

#> $id
#> [1] "users/csaybar/rgee/rgee_folder/rgee_ic_moved"
#> 
#> $version
#> [1] 1880803599
#> 
#> $features
#> list()
#> 
#> $type
#> [1] "ImageCollection"
#> 
#> $bands
#> list()
#> 

# 8. Share EE objects -- Create a public dataset
ee_manage_assets_access('users/csaybar/rgee/rgee_folder/rgee_ic_moved')

#> The ACL of users/csaybar/rgee/rgee_folder/rgee_ic_moved has been changed.

ee$data$getAssetAcl('users/csaybar/rgee/rgee_folder/rgee_ic_moved')

#> $owners
#> [1] "csaybar@gmail.com"
#> 
#> $readers
#> [1] "csaybar@gmail.com"
#> 
#> $all_users_can_read
#> [1] TRUE
#> 
#> $writers
#> [1] "csaybar@gmail.com"
#> 

# 9. Create a report based on all tasks that is running or has finished
ee_manage_task()

#>                          ID     State
#> 1  PTVMRXDVLFIN2SZBK6BGJXHW COMPLETED
#> 2  MJFZIMQHV36M4KSJYJWBMHPU COMPLETED
#> 3  TWNCBPOU5BZVXVXSOKRG2EOB COMPLETED
#> 4  MV4WJRHX36H4ZIOSYB4B2EEK    FAILED
#> 5  XZZXVPZOMRBHJURCHWLNXZIR COMPLETED
#> 6  SHLX6ZAMXESKP72MBFJGDS4O COMPLETED
#> 7  H5UO5VZTQWUA7ITC62Z2NIE5    FAILED
#> 8  MCVBEUWB55TSEPMA335NXMBT COMPLETED
#> 9  TTNEWGY62MA5GL6ZYX4YD747 COMPLETED
#> 10 DZJXKUAI2RHJU7SZ35RCFWRI COMPLETED
#> 11 JVTVMLH5S6A7OVTLVQEG7PDZ COMPLETED
#> 12 KKPCN4MYGKS76RQYHQ3RMUQX COMPLETED
#> 13 2XMHMTDA7FENFGQ5JCH6N5BC COMPLETED
#> 14 YZJATEQVYFTCPGPOKQZFR6A2 COMPLETED
#> 15 FREOXBCBXVY45LS6LXJRJAGP COMPLETED
#>                               DestinationPath            Type
#> 1                           myExportImageTask    EXPORT_IMAGE
#> 2         users/csaybar/rgee_upload/geomatrix          INGEST
#> 3          users/csaybar/rgee_upload/L7_stars          INGEST
#> 4  users/csaybar/rgee_upload/arequipaL7_stars          INGEST
#> 5  users/csaybar/rgee_upload/arequipaL7_stars          INGEST
#> 6  users/csaybar/rgee_upload/arequipaarequipa          INGEST
#> 7                   users/csaybar/rgee_upload          INGEST
#> 8                           myExportTableTask EXPORT_FEATURES
#> 9                           myExportImageTask    EXPORT_IMAGE
#> 10                    users/csaybar/L7_ETM22s          INGEST
#> 11                         users/csaybar/xxxx          INGEST
#> 12                     users/csaybar/L7_ETMs2          INGEST
#> 13             users/csaybar/file6ef723f4ae4b          INGEST
#> 14             users/csaybar/file6ef76af6cb1e          INGEST
#> 15                      users/csaybar/L7_ETMs          INGEST
#>                  Start DeltaToCreate(s) DeltaToCompletedTask(s)
#> 1  2019-09-10 19:25:30 5.78099989891052         9.0789999961853
#> 2  2019-09-10 06:04:03 6.64199995994568        23.8129999637604
#> 3  2019-09-10 05:54:49 4.40499997138977        20.1119999885559
#> 4  2019-09-10 05:37:12 4.47299981117249       0.705000162124634
#> 5  2019-09-10 05:35:11 6.19799995422363        23.6330001354218
#> 6  2019-09-10 05:08:52 4.90100002288818        544.118000030518
#> 7  2019-09-10 04:58:29 5.99699997901917       0.569000005722046
#> 8  2019-09-09 19:02:11 7.44099998474121        5.93299984931946
#> 9  2019-09-09 19:01:12 3.18499994277954        9.25399994850159
#> 10 2019-09-01 02:44:57 6.91200017929077        25.4869999885559
#> 11 2019-09-01 02:42:17 5.72699999809265        19.1160001754761
#> 12 2019-08-31 23:12:38 6.58400011062622        26.0720000267029
#> 13 2019-08-31 20:59:57 5.68899989128113        23.2569999694824
#> 14 2019-08-31 20:29:13 11.2030000686646        23.6689999103546
#> 15 2019-08-31 20:13:31 4.08299994468689         29.055999994278
#>                                                            ErrorMessage
#> 1                                                                  NULL
#> 2                                                                  NULL
#> 3                                                                  NULL
#> 4  Cannot overwrite asset 'users/csaybar/rgee_upload/arequipaL7_stars'.
#> 5                                                                  NULL
#> 6                                                                  NULL
#> 7                   Cannot overwrite asset 'users/csaybar/rgee_upload'.
#> 8                                                                  NULL
#> 9                                                                  NULL
#> 10                                                                 NULL
#> 11                                                                 NULL
#> 12                                                                 NULL
#> 13                                                                 NULL
#> 14                                                                 NULL
#> 15                                                                 NULL

# 10. Cancel all the running task
ee_manage_cancel_all_running_taks()

# 11. Delete EE objects or folders
ee_manage_delete('users/csaybar/rgee/')

#> EE object deleted: users/csaybar/rgee/rgee_folder/rgee_ic_moved 
#> EE object deleted: users/csaybar/rgee/rgee_folder 
#> EE object deleted: users/csaybar/rgee