Print and return metadata about Spatial Earth Engine Objects

Print and return metadata about Spatial Earth Engine Objects

ee_print(eeobject, max_display = 10)

# S3 method for ee.geometry.Geometry
ee_print(eeobject, max_display = 10)

# S3 method for ee.feature.Feature
ee_print(eeobject, max_display = 10)

# S3 method for ee.featurecollection.FeatureCollection
ee_print(eeobject,
  max_display = 10)

# S3 method for ee.image.Image
ee_print(eeobject, max_display = 10)

# S3 method for ee.imagecollection.ImageCollection
ee_print(eeobject,
  max_display = 10)

Arguments

eeobject
Earth Engine Object. Available for: Geometry, Feature, FeatureCollection, Image or ImageCollection.

max_display
Set the max number of properties to display.

Details

For avoiding computing time extremely large and "EEException" related with memory consuming (See this Earth Engine Guide) just the first element of: ee$Image, ee$ImageCollection, ee$Feature and ee$FeatureCollection will be used to generated Earth Engine object properties.

Examples

library(rgee)
ee_Initialize()

#> Requesting Earth Engine authorization  ... DONE

eeobject <- ee$ImageCollection("LANDSAT/LC08/C01/T1_TOA")$
  filter(ee$Filter()$eq("WRS_PATH", 44))$
  filter(ee$Filter()$eq("WRS_ROW", 34))$
  filterDate("2014-03-01", "2014-08-01")
ee_print(eeobject, max_display = 118)

#> Class                          : ImageCollection 
#> nImages                        : 9 
#> nBands                         : 12 
#> Bands names                    : B1 B2 B3 B4 B5 B6 B7 B8 B9 B10 B11 BQA 
#> EPSG (SRID)                    : 32610 
#> proj4string                    : +proj=utm +zone=10 +datum=WGS84 +units=m +no_defs 
#> Geotransform                   : [30 0 460785 0 -30 4264215] 
#> Scale (m)                      : 30 
#> Dimensions                     : nrow:7661  ncol:7801 
#> nPixels by image               : 59'763'461 
#> Data type                      : FLOAT 
#> Image Properties               : 118 
#>                                |        RADIANCE_ADD_BAND_11        |       REFLECTANCE_ADD_BAND_9       |
#>                                |        RADIANCE_ADD_BAND_6         |      REFLECTANCE_MULT_BAND_7       |
#>                                |       GEOMETRIC_RMSE_MODEL_Y       |       REFLECTANCE_ADD_BAND_5       |
#>                                |              UTM_ZONE              |        RADIANCE_ADD_BAND_8         |
#>                                |       GEOMETRIC_RMSE_MODEL_X       |           MAP_PROJECTION           |
#>                                |        K1_CONSTANT_BAND_11         |      REFLECTANCE_MULT_BAND_9       |
#>                                |      REFLECTANCE_MULT_BAND_4       |              CPF_NAME              |
#>                                |          LANDSAT_SCENE_ID          |        RADIANCE_ADD_BAND_5         |
#>                                |        COLLECTION_CATEGORY         |           DATE_ACQUIRED            |
#>                                |       REFLECTANCE_ADD_BAND_8       |        K2_CONSTANT_BAND_11         |
#>                                |        RADIANCE_ADD_BAND_2         |          ELEVATION_SOURCE          |
#>                                |        RADIANCE_MULT_BAND_6        |    google:registration_offset_y    |
#>                                |        RADIANCE_MULT_BAND_3        |         COLLECTION_NUMBER          |
#>                                |         IMAGE_QUALITY_OLI          |       REFLECTANCE_ADD_BAND_1       |
#>                                |        RADIANCE_ADD_BAND_3         |        RADIANCE_MULT_BAND_7        |
#>                                |              WRS_ROW               |         SATURATION_BAND_6          |
#>                                |      REFLECTANCE_MULT_BAND_3       |        RADIANCE_ADD_BAND_7         |
#>                                |          THERMAL_SAMPLES           |         SATURATION_BAND_5          |
#>                                |           THERMAL_LINES            |     GRID_CELL_SIZE_REFLECTIVE      |
#>                                |           BPF_NAME_TIRS            |          REFLECTIVE_LINES          |
#>                                |      TIRS_SSM_POSITION_STATUS      |        GEOMETRIC_RMSE_MODEL        |
#>                                |            BPF_NAME_OLI            |        RADIANCE_ADD_BAND_1         |
#>                                |      REFLECTANCE_MULT_BAND_8       |      REFLECTANCE_MULT_BAND_1       |
#>                                |        RADIANCE_MULT_BAND_9        |           SUN_ELEVATION            |
#>                                |         PANCHROMATIC_LINES         |         SATURATION_BAND_8          |
#>                                |       REFLECTANCE_ADD_BAND_6       |             REQUEST_ID             |
#>                                |       REFLECTANCE_ADD_BAND_2       |        PANCHROMATIC_SAMPLES        |
#>                                |       REFLECTANCE_ADD_BAND_7       |         REFLECTIVE_SAMPLES         |
#>                                |           NADIR_OFFNADIR           |   GROUND_CONTROL_POINTS_VERSION    |
#>                                |        RADIANCE_MULT_BAND_8        |    GROUND_CONTROL_POINTS_MODEL     |
#>                                |    GROUND_CONTROL_POINTS_VERIFY    |         SATURATION_BAND_1          |
#>                                |             DATA_TYPE              |        RADIANCE_MULT_BAND_4        |
#>                                |        RADIANCE_ADD_BAND_9         |    GRID_CELL_SIZE_PANCHROMATIC     |
#>                                |         LANDSAT_PRODUCT_ID         |          TARGET_WRS_PATH           |
#>                                |           RLUT_FILE_NAME           |           TRUNCATION_OLI           |
#>                                |             STATION_ID             |              WRS_PATH              |
#>                                |     google:registration_count      |         SATURATION_BAND_9          |
#>                                |           TARGET_WRS_ROW           |        RADIANCE_ADD_BAND_4         |
#>                                |        RADIANCE_MULT_BAND_1        |           SPACECRAFT_ID            |
#>                                |         RESAMPLING_OPTION          |             ELLIPSOID              |
#>                                |          system:footprint          |             ROLL_ANGLE             |
#>                                |       REFLECTANCE_ADD_BAND_4       |     google:registration_ratio      |
#>                                |            ORIENTATION             |        K2_CONSTANT_BAND_10         |
#>                                |         system:asset_size          |       RADIANCE_MULT_BAND_10        |
#>                                |    PROCESSING_SOFTWARE_VERSION     |      REFLECTANCE_MULT_BAND_6       |
#>                                |            CLOUD_COVER             |           TIRS_SSM_MODEL           |
#>                                | TIRS_STRAY_LIGHT_CORRECTION_SOURCE |        RADIANCE_MULT_BAND_5        |
#>                                |       RADIANCE_MULT_BAND_11        |            system:index            |
#>                                |         SATURATION_BAND_2          |         system:time_start          |
#>                                |         EARTH_SUN_DISTANCE         |            SUN_AZIMUTH             |
#>                                |        RADIANCE_ADD_BAND_10        |       GRID_CELL_SIZE_THERMAL       |
#>                                |    google:registration_offset_x    |        RADIANCE_MULT_BAND_2        |
#>                                |         SATURATION_BAND_3          |      REFLECTANCE_MULT_BAND_5       |
#>                                |         SATURATION_BAND_7          |         SATURATION_BAND_4          |
#>                                |       REFLECTANCE_ADD_BAND_3       |               DATUM                |
#>                                |       GEOMETRIC_RMSE_VERIFY        |          CLOUD_COVER_LAND          |
#>                                |        K1_CONSTANT_BAND_10         |      REFLECTANCE_MULT_BAND_2       |
#>                                |             FILE_DATE              |         SCENE_CENTER_TIME          |
#>                                |             SENSOR_ID              |         IMAGE_QUALITY_TIRS         |
#> ImageCollection Properties     : 26 
#>                                |           provider           |            thumb             |
#>                                |  system:visualization_1_max  | system:visualization_1_bands |
#>                                | system:visualization_1_name  | system:visualization_2_name  |
#>                                |  system:visualization_2_min  | system:visualization_0_gain  |
#>                                |            sample            |          date_range          |
#>                                |             tags             |         product_tags         |
#>                                |         description          | system:visualization_1_gain  |
#>                                |            period            |         source_tags          |
#>                                |         provider_url         | system:visualization_2_bands |
#>                                |  system:visualization_0_max  |            title             |
#>                                | system:visualization_2_gain  | system:visualization_0_bands |
#>                                |  system:visualization_0_min  | system:visualization_0_name  |
#>                                |  system:visualization_2_max  |  system:visualization_1_min  |