Wrap an R function in a Python function with the same signature.

This function could wrap an R function in a Python function with the same signature. Note that the signature of the R function must not contain esoteric Python-incompatible constructs.

ee_pyfunc(f)

Arguments

f
An R function

Value

A Python function that calls the R function f with the same signature.

Note

py_func has been renamed to ee_pyfunc just to maintain the rgee functions name's style. All recognition for this function must always be given to reticulate.

Examples

library(rgee)
ee_reattach() # reattach ee as a reserved word
ee_Initialize()

#> ── rgee 0.3.10 ───────────────────────────────────────────────────────────────────────── earthengine-api 0.1.217 ──

#>  ✓ Python version: /home/aybarpc01/.virtualenvs/ee/bin/python3 
#>  ✓ email: not_defined
#>  ✓ Initializing Google Earth Engine: ✓ Initializing Google Earth Engine:  DONE!
#>  ✓ Earth Engine user: users/datacolecfbf 

#> ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────

# Earth Engine List
ee_SimpleList <- ee$List$sequence(0, 12)
ee_NewList <- ee_SimpleList$map(
  ee_pyfunc(
    function(x) {
      ee$Number(x)$add(x)
    }
  )
)
ee_NewList$getInfo()

#>  [1]  0  2  4  6  8 10 12 14 16 18 20 22 24

# Earth Engine ImageCollection
constant1 <- ee$Image(1)
constant2 <- ee$Image(2)
ee_ic <- ee$ImageCollection(c(constant2, constant1))
ee_newic <- ee_ic$map(
  ee_pyfunc(
    function(x) ee$Image(x)$add(x)
  )
)
ee_newic$mean()$getInfo()$type

#> [1] "Image"