Interface to search into the Earth Engine Data Catalog

R functions for searching in Earth Engine's public data archive.

ee_dataset(quiet = FALSE, upgrade = FALSE)

ee_search_startdate(ee_dataset, stardate)

ee_search_enddate(ee_dataset, enddate = Sys.Date())

ee_search_type(ee_dataset, type)

ee_search_provider(ee_dataset, provider)

ee_search_provider_list(ee_dataset)

ee_search_tags(ee_dataset, ..., logical_operator = "OR")

ee_search_title(ee_dataset, ..., logical_operator = "OR")

ee_search_tagstitle(ee_dataset, ..., logical_operator = "OR")

ee_search_title_list(ee_dataset)

ee_search_display(ee_dataset, maxdisplay = 10)

Arguments

quiet
logical. Suppress info message

upgrade
Logical. If the dataset needs to be upgraded.

ee_dataset
data.frame. Generated by rgee::ee_search_Datasets(). It is automatically updated and expanded weekly.

stardate
Character. Start date of dataset availability.

enddate
Character. End date of dataset availability.

type
Character. "Image", "ImageCollection" or a "table".

provider
Character. Name of the dataset's provider. See ee_search_provider_list()

...
Character vector. tags

logical_operator
Character. Available just for rgee::ee_search_tags and rgee::ee_search_title. 'AND' represents inclusiveness between tags in searching and 'OR' exclusiveness.

maxdisplay
Numeric. Maximum number of tabs to display in their browser

Value

a data.frame.

Examples

library(rgee)

ee_reattach() # reattach ee as a reserved word

#> Error in py_module_import(module, convert = convert): ImportError: No module named ee

ee_Initialize()

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#> 

# ee_search_provider_list()
# ee_search_title_list()
myquery <- ee_dataset() %>%
  ee_search_type("Image") %>%
  ee_search_provider("WWF") %>%
  ee_search_tags("srtm", "flow", "direction", "dem") %>%
  ee_search_title("15", "Flow", logical_operator = "AND")

#> Downloading(Upgrading) the Earth Engine catalog ... please wait

#> Warning: cannot open file '/Users/runner/.config/earthengine/ee_dataset.csv': No such file or directory

#> Error in file(file, ifelse(append, "a", "w")): cannot open the connection

print(myquery$id)

#> Error in print(myquery$id): object 'myquery' not found