Upload sf or stars objects into a GEE asset

Upload either images or vectors to Google Earth Engine asset

ee_upload(x, ...)

# S3 method for character
ee_upload(x, ..., filename, bucket = NULL,
  properties = NULL, start_time = "1970-01-01",
  end_time = "1970-01-01",
  selenium_params = getOption("rgee.selenium.params"), clean = FALSE,
  reinit = FALSE, quiet = FALSE)

# S3 method for sf
ee_upload(x, ..., filename, bucket = NULL,
  selenium_params = getOption("rgee.selenium.params"), clean = FALSE,
  reinit = FALSE, quiet = FALSE)

# S3 method for stars
ee_upload(x, ..., filename, bucket = NULL,
  properties = NULL, start_time = "1970-01-01",
  end_time = "1970-01-01",
  selenium_params = getOption("rgee.selenium.params"), clean = FALSE,
  reinit = FALSE, quiet = FALSE)

# S3 method for stars_proxy
ee_upload(x, ..., filename, bucket = NULL,
  properties = NULL, start_time = "1970-01-01",
  end_time = "1970-01-01",
  selenium_params = getOption("rgee.selenium.params"), clean = FALSE,
  reinit = FALSE, quiet = FALSE)

Arguments

x
Character, sf or stars object to upload into a GEE asset.

...
ignored

filename
Character. Asset destination path, e.g. users/pinkiepie/myponycollection.

bucket
bucketname you are uploading to. See details.

properties
List. Set of parameters to established as a property of an EE object. See details.

start_time
Character. The timestamp associated with the asset. The initial time stamp is set to the nominal image acquisition time for single scenes.

end_time
Character. Useful for assets that correspond to an interval of time. The ending time stamp is set to the nominal image acquisition time for single scenes.

selenium_params
List. Optional parameters when bucket is NULL. Parameters for setting selenium. See details.

clean
Logical; whether is TRUE cache will cleaned, see Description.

reinit
Logical; run ee_Initialize(gcs=TRUE) before start to upload

quiet
Logical. Suppress info message.

Details

It is necessary, for uploading process, get authorization to read & write into a Google Cloud Storage (GCS) bucket. Earth Engine provides a provisional for free space into GCS through gs://earthengine-uploads/. If the bucket argument is absent, this function will use Selenium driver for getting access to the URI mentioned bellow, see ee_upload_file_to_gcs for details. Install and check the Selenium drivers for Google Chrome is possible as follow:

rgee::ee_install_drivers()

rgee::ee_check_drivers()

The properties argument is just available for image uploads, is you are interesting in setting properties in FeatureCollection. Please use ee_manage_set_properties after ee_upload.

The selenium_params argument is a three-element list consisting of:

gmail_account: The google account. If it is not specified, it will obtained from ee$data$getAssetRoots().

showpassword: Logical. After put the google account into getPass, should be shown?.

cache: Logical. TRUE will use the cookies saved on the /temp directory.

With respect to the variables time_start and time_end, both needs to be specified as seconds since the epoch (1970-01-01). Assumed to be in the UTC time zone.

Examples

library(rgee)
library(stars)
library(sf)

username <- 'aybar1994' #change according to username.
gcs_bucket <- 'bag_csaybar'
ee_check_drivers()

#> 
#> ◉  Selenium drivers: 
#> 
#> ✔ [Ok] Chromedriver 
#> 

ee_Initialize(user_gmail = username, gcs = TRUE)

#> user_gmail: aybar1994 
#> assethome: users/aybar1994 
#> Checking Google Cloud storage credentials ...  DONE
#> Checking correspondence between user_gmail and assethome ... DONE 

# Create a folder in Earth Engine Asset
filename <- sprintf("users/%s/rgee_upload/", username)
ee_manage_create(filename)

#> GEE asset: users/aybar1994/rgee_upload created

# Select image to upload
tif = system.file("tif/geomatrix.tif", package = "stars")
geomatrix = read_stars(tif) %>% st_warp(crs=st_crs(4326))

# Upload to earth egnine
ee_upload(x = geomatrix,
          filename = paste0(filename,"geomatrix"),
          bucket = gcs_bucket)

#> Set default bucket name to 'bag_csaybar'

#> 2020-01-19 22:29:57 -- File size detected as 2.9 Kb

# Read uploaded image
asset_geomatrix <- paste0(filename,"geomatrix")
ee_geomatrix <- ee$Image(asset_geomatrix)
ee_map(ee_geomatrix, zoom_start = 18)

#> Error in py_call_impl(callable, dots$args, dots$keywords): EEException: Image.load: Image asset 'users/aybar1994/rgee_upload/geomatrix' not found.
#> 
#> Detailed traceback: 
#>   File "/home/aybarpc01/.local/lib/python3.7/site-packages/ee/image.py", line 118, in getInfo
#>     return super(Image, self).getInfo()
#>   File "/home/aybarpc01/.local/lib/python3.7/site-packages/ee/computedobject.py", line 95, in getInfo
#>     return data.computeValue(self)
#>   File "/home/aybarpc01/.local/lib/python3.7/site-packages/ee/data.py", line 717, in computeValue
#>     prettyPrint=False))['result']
#>   File "/home/aybarpc01/.local/lib/python3.7/site-packages/ee/data.py", line 340, in _execute_cloud_call
#>     raise _translate_cloud_exception(e)

## OPTIONAL: add properties
ee_manage_set_properties(
  path_asset = asset_geomatrix,
  properties = list(message='hello-world',language = 'R'))

#> Error in ee_manage_set_properties(path_asset = asset_geomatrix, properties = list(message = "hello-world",     language = "R")): unused argument (properties = list(message = "hello-world", language = "R"))

# Clean EE asset and GCS
ee_manage_delete(dirname(asset_geomatrix))

#> EE object deleted: users/aybar1994/rgee_upload 

googleCloudStorageR::gcs_global_bucket(gcs_bucket)

#> Set default bucket name to 'bag_csaybar'

buckets <- googleCloudStorageR::gcs_list_objects()
mapply(googleCloudStorageR::gcs_delete_object, buckets$name)

#> geomatrix.tif 
#>          TRUE