Upload sf or stars objects to GEE asset

Send either images or vectors to Google Earth Engine asset through Selenium

ee_upload_toasset(x, ...)

# S3 method for character
ee_upload_toasset(x, ..., gmail_account,
  gee_asset_path, properties = NULL, pyramiding_policy = "MEAN",
  one_by_one = TRUE, quiet = FALSE)

# S3 method for sf
ee_upload_toasset(x, ..., gmail_account, asset_filename,
  gee_asset_path, quiet = FALSE)

# S3 method for stars
ee_upload_toasset(x, ..., gmail_account, gee_asset_path,
  type = "Float32", properties = NULL, pyramiding_policy = "MEAN",
  one_by_one = TRUE, quiet = FALSE)

# S3 method for stars_proxy
ee_upload_toasset(x, ..., gmail_account,
  gee_asset_path, properties = NULL, pyramiding_policy = "MEAN",
  one_by_one = TRUE, quiet = FALSE)

Arguments

x
sf or stars object.

...
ignored

gmail_account
character; google account name (gmail address).

gee_asset_path
character; images destination path, full path for upload to Google Earth Engine, e.g. users/pinkiepie/myponycollection.

properties
data.frame; Manifest upload. See details

pyramiding_policy
character; determines how each pixel at a given level of the pyramid is computed from the aggregation. "MEAN" (default), "MODE" and "SAMPLE".

one_by_one
TODO

quiet
logical; suppress info message

asset_filename
asset filename

type
TODO

Examples

# NOT RUN {
library(rgee)
# }

Developed by Roy Yali, Cesar Aybar.

Site built with pkgdown 1.3.0.