Create a manifest to upload an image

Create a manifest to upload a GeoTIFF to Earth Engine assets folder. The "manifest" is simply a JSON file which describe all the upload parameters. See https://developers.google.com/earth-engine/image_manifest to get more details.

ee_utils_create_manifest_image(
  gs_uri,
  assetId,
  properties = NULL,
  start_time = "1970-01-01",
  end_time = "1970-01-01",
  pyramiding_policy = "MEAN",
  returnList = FALSE,
  quiet = FALSE
)

Arguments

gs_uri
Character. GCS full path of the image to upload to Earth Engine assets e.g. gs://rgee_dev/l8.tif

assetId
Character. How to call the file once uploaded to the Earth Engine Asset. e.g. users/datacolecfbf/l8.

properties
List. Set of parameters to be set up as properties of the EE object.

start_time
Character. Sets the start time property (system:time_start) It could be a number (timestamp) or a date.

end_time
Character. Sets the end time property (system:time_end) It could be a number (timestamp) or a date.

pyramiding_policy
Character. The pyramid reduction policy to use.

returnList
Logical. If TRUE will return the "manifest" as a list otherwise will return a JSON file.

quiet
Logical. Suppress info message.

Value

If returnList is TRUE a list otherwise a JSON file.

See also

Examples

if (FALSE) {
library(rgee)
ee_Initialize()

tif <- system.file("tif/L7_ETMs.tif", package = "stars")

# Return a JSON file
ee_utils_create_manifest_image(
  gs_uri = "gs://rgee_dev/l8.tif",
  assetId = "users/datacolecfbf/l8"
)

# Return a list
ee_utils_create_manifest_image(
  gs_uri = "gs://rgee_dev/l8.tif",
  assetId = "users/datacolecfbf/l8",
  returnList = TRUE
)
}