The value of a future or the values of all elements in a container

Gets the value of a future or the values of all elements (including futures) in a container such as a list, an environment, or a list environment. If one or more futures is unresolved, then this function blocks until all queried futures are resolved.

ee_utils_future_value(future, stdout = TRUE, signal = TRUE, ...)

Arguments

future,
x A Future, an environment, a list, or a list environment.

stdout
If TRUE, standard output captured while resolving futures is relayed, otherwise not.

signal
If TRUE, conditions captured while resolving futures are relayed, otherwise not.

...
All arguments used by the S3 methods.

Value

value() of a Future object returns the value of the future, which can be any type of R object.

value() of a list, an environment, or a list environment returns an object with the same number of elements and of the same class. Names and dimension attributes are preserved, if available. All future elements are replaced by their corresponding value() values. For all other elements, the existing object is kept as-is.

If signal is TRUE and one of the futures produces an error, then that error is produced.

Author

Henrik Bengtsson https://github.com/HenrikBengtsson/