Pass an Earth Engine date object to R

Pass an Earth Engine date object to R

eedate_to_rdate(ee_date, timestamp = FALSE)

Arguments

ee_date
ee$date object (ee$Date)

timestamp
Logical. If TRUE, return the date in milliseconds from the Unix Epoch (1970-01-01 00:00:00 UTC) otherwise return the date as a POSIXct object. By default FALSE.

Value

eedate_to_rdate will return either a numeric timestamp or a POSIXct object depending on the timestamp argument.

Details

eedate_to_rdate is essential to avoid potential errors that might appear when users call to retrieve dates. Currently, R integer only support 32 bit signed, such integers can only count up to about 2 billion. This range is extremely insufficient for dealing with Google Earth Engine date objects (ee$Date), which are represented by timestamps in milliseconds since the UNIX epoch. eedate_to_rdate use Python as a backend to obtain the date and convert it in float before exporting to R.

See also

Examples

if (FALSE) {
library(rgee)
ee_Initialize()

eeDate <- ee$Date$fromYMD(2010,1,1)
eedate_to_rdate(eeDate,timestamp = TRUE) # good
eeDate$getInfo()$value # bad
}