Pass an R date object to Earth Engine

Pass an R date object ("Date", "Numeric", "character", "POSIXt", and "POSIXct") to Google Earth Engine (ee$Date).

rdate_to_eedate(date, timestamp = FALSE)

Arguments

date
R date object

timestamp
Logical. If TRUE, return the date in milliseconds from the Unix Epoch (1970-01-01 00:00:00 UTC) otherwise return a EE date object. By default FALSE.

Value

rdate_to_eedate will return a numeric timestamp or an EE Date depending on the timestamp argument.

Examples

library(rgee)
ee_reattach()

#> Error in py_module_import(module, convert = convert): ImportError: No module named ee

ee_Initialize()

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#> 

rdate_to_eedate('2000-01-01')

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#> 

rdate_to_eedate(315532800000) # float number

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#>