rgee: R Bindings for Calling the 'Earth Engine' API

Earth Engine <https://earthengine.google.com/> client library for R. All of the 'Earth Engine' API classes, modules, and functions are made available. Additional functions implemented include importing (exporting) of Earth Engine spatial objects, extraction of time series, interactive map display, assets management interface, and metadata display. See <https://r-spatial.github.io/rgee/> for further details.

See also

Author

Maintainer: Cesar Aybar csaybar@gmail.com (ORCID)

Other contributors:

Wu Qiusheng giswqs@gmail.com (ORCID) [contributor]

Lesly Bautista leslyarcelly.213@gmail.com (ORCID) [contributor]

Roy Yali ryali93@gmail.com (ORCID) [contributor]

Antony Barja antony.barja8@gmail.com (ORCID) [contributor]

Kevin Ushey kevin@rstudio.com [contributor]

Jeroen Ooms jeroen@berkeley.edu (ORCID) [contributor]

Tim Appelhans tim.appelhans@gmail.com [contributor]

JJ Allaire jj@rstudio.com [contributor]

Yuan Tang terrytangyuan@gmail.com [contributor]

Samapriya Roy samapriya.roy@gmail.com [contributor]

MariaElena Adauto 2a.mariaelena@gmail.com (ORCID) [contributor]

Gabriel Carrasco gabriel.carrasco@upch.pe (ORCID) [contributor]

Henrik Bengtsson henrikb@braju.com [contributor]

Jeffrey Hollister hollister.jeff@epa.gov (Hollister reviewed the package for JOSS, see https://github.com/openjournals/joss-reviews/issues/2272/) [reviewer]

Gennadii Donchyts (Gena reviewed the package for JOSS, see https://github.com/openjournals/joss-reviews/issues/2272/) [reviewer]

Marius Appel marius.appel@uni-muenster.de (Appel reviewed the package for JOSS, see https://github.com/openjournals/joss-reviews/issues/2272/) [reviewer]