Convert R spatial object to an Google Earth Engine

Convert R spatial object to an Google Earth Engine

sf_as_ee(x, ...)

ee_as_ee.character(x, ...)

ee_as_ee.sf(x, ...)

ee_as_ee.sfc(x, ...)

ee_as_ee.sfg(x, ...)

Arguments

x
object to be converted into.

...
if x is a character additional st_read arguments could be passed.

Details

sf_as_ee try to transform R objects (sf) into a GeoJSON format. By the time a user sends an HTTP request to the Earth Engine Web REST APIs (by, e.g. using *$getInfo()), the GeoJSON will be encrusted to the message. Therefore, it is expected that large spatial objects (>500Kb) cause bottlenecks and plodding connections. See Client vs Server documentation for details. For leading with very large spatial objects, is a good practice firstly importing it to the GEE assets. See ee_upload_toasset for details.

Examples

# NOT RUN {
library(rgee)
ee_initialize()

# character
x <- system.file("shape/nc.shp", package="sf")
ee$FeatureCollection(ee_as_ee(x),proj="EPSG:4326")

# matrix
ee_df <- matrix(rnorm(100),ncol = 5) %>% as.data.frame()
ee_df <- ee_as_ee(ee_df,c("V1","V2"))

ee_df$geometry()
# sf
x <- read_sf(system.file("shape/nc.shp", package="sf"))
st_crs
ee_as_ee(x)

#sfc
x <- read_sf(system.file("shape/nc.shp", package="sf"))$geometry
ee_as_ee(x)

#sfg
x <- read_sf(system.file("shape/nc.shp", package="sf"))$geometry[[1]]
ee_as_ee(x)

#sp
x <- shapefile(system.file("shape/nc.shp", package="sf"))
ee_as_ee(x)
# }

Developed by Roy Yali, Cesar Aybar.

Site built with pkgdown 1.3.0.