Authenticate and Initialize Earth Engine

Authorize rgee to manage Earth Engine resources, Google Drive, and Google Cloud Storage. The ee_initialize() via web-browser will ask to sign in to your Google account and allows you to grant permission to manage resources. This function is a wrapper around rgee::ee$Initialize().

ee_Initialize(
  email = NULL,
  drive = FALSE,
  gcs = FALSE,
  display = FALSE,
  quiet = FALSE
)

Arguments

email
Character (optional, e.g. data.colec.fbf@gmail.com). The email argument is used to create a folder inside the path ~/.config/earthengine/ that save all credentials for a specific Google identity.

drive
Logical (optional). If TRUE, the drive credential will be cached in the path ~/.config/earthengine/.

gcs
Logical (optional). If TRUE, the Google Cloud Storage credential will be cached in the path ~/.config/earthengine/.

display
Logical. If TRUE display the earthengine authentication URL.

quiet
Logical. Suppress info messages.

Value

No return value, called for initializing the earthengine-api.

Details

ee_Initialize(...) can manage Google drive and Google Cloud Storage resources using the R packages googledrive and googlecloudStorageR, respectively. By default, rgee does not require them, these are only necessary to enable rgee I/O functionality. All user credentials are saved in the directory ~/.config/earthengine/, if a user does not specify the email argument all user credentials will be saved in a subdirectory called ~/.config/earthengine/ndef.

See also

Examples

if (FALSE) {
library(rgee)

# Simple init - Load just the Earth Engine credential
ee_Initialize()
ee_user_info()
}