Authenticate and Initialize Earth Engine

Authorize rgee to manage Earth Engine resources, Google Drive, and Google Cloud Storage. The ee_initialize() via web-browser will ask users to sign into your Google account and allows you to grant permission to manage resources. This function is a wrapper around rgee::ee$Initialize().

ee_Initialize(
  user = NULL,
  drive = FALSE,
  gcs = FALSE,
  credentials = "persistent",
  opt_url = NULL,
  cloud_api_key = NULL,
  http_transport = NULL,
  project = NULL,
  quiet = FALSE,
  auth_mode = "notebook",
  auth_quiet = FALSE,
  ...
)

Arguments

user

Character (optional, e.g. data.colec.fbf). The user parameter is used to create a folder inside the path ~/.config/earthengine/ where all the credentials for a specific Google identity are saved.

drive

Logical (optional). If set to TRUE, the drive credential will be cached in the path ~/.config/earthengine/.

gcs

Logical (optional). If TRUE, the Google Cloud Storage credential will be cached in the path ~/.config/earthengine/.

credentials

OAuth2 GEE credentials. 'persistent' (default) means it will use the GEE credentials already stored in the filesystem. If the credentials are not found, it will raise an explanatory exception guiding the user to create those credentials.

opt_url

The base url for the EarthEngine REST API to connect to.

cloud_api_key

An optional API key to use the Cloud API.

http_transport

The HTTP transport method to use for making requests

project

The client project ID or number to be used for making API calls.

quiet

Logical. Suppress info messages.

auth_mode

The authentication mode. One of:

1. paste - send user to accounts.google.com to get a pastable token

2. notebook - send user to notebook authenticator page

3. gcloud - use gcloud to obtain credentials (will set appdefault)

4. appdefault - read from existing $GOOGLE_APPLICATION_CREDENTIALS file

5. None - a default mode is chosen based on your environment.

auth_quiet

Logical. ee_Authenticate quiet parameter. If TRUE, do not require interactive prompts and force –no-browser mode for gcloud.

...

Extra exporting argument. See ee_Authenticate.

Value

No return value, called for initializing the earthengine-api.

Details

ee_Initialize() can manage Google Drive, and Google Cloud Storage resources using the R packages googledrive and googlecloudStorageR, respectively. By default, rgee does not require them. These are only necessary to enable rgee I/O functionality. All user credentials are saved in the directory ~/.config/earthengine/.

See also

Examples

if (FALSE) { # \dontrun{
library(rgee)

# Simple init - Load just the Earth Engine credential
ee_Initialize()
ee_user_info()
} # }