Create an isolated Python virtual environment with all rgee dependencies.

Create an isolated Python virtual environment with all rgee dependencies. ee_install realize the following six (6) tasks:

1. If you do not count with a Python environment, it will display an interactive menu to install Miniconda (a free minimal installer for conda).

2. Remove the previous Python environment defined in py_env if it exist.

3. Create a new Python environment (See py_env).

4. Set the environmental variable EARTHENGINE_PYTHON. It is used to define RETICULATE_PYTHON when the library is loaded. See this article for further details.

5. Install rgee Python dependencies. Using reticulate::py_install.

6. Interactive menu to confirm if restart the R session to see changes.

ee_install(
  py_env = "rgee",
  earthengine_version = ee_version(),
  confirm = interactive()
)

Arguments

py_env
Character. The name, or full path, of the Python environment to be used by rgee.

earthengine_version
Character. The Earth Engine Python API version to install. By default rgee::ee_version().

confirm
Logical. Confirm before restarting R?.

See also

Examples

if (FALSE) {
# library(rgee)
# ee_install() #It is just necessary once
}