Create an isolated Python virtual environment with all rgee dependencies.

Create an isolated Python virtual environment with all rgee dependencies. ee_install realize the following six (6) tasks:

1. IIf you do not have a Python environment installed, it will display an interactive menu to install Miniconda (a free minimal installer for conda).

2. If it exists, the previous Python environment specified in the py_env argument will be deleted.

3. Create a new Python environment (See py_env) argument.

4. Set the environment variable EARTHENGINE_PYTHON and EARTHENGINE_ENV. It is used to define RETICULATE_PYTHON when the library is loaded. See this article for further details.

5. Install rgee Python dependencies. Using reticulate::py_install.

6. Interactive menu to confirm if restart the R session to see changes.

ee_install(
  py_env = "rgee",
  earthengine_version = ee_version(),
  python_version = "3.8",
  confirm = interactive()
)

Arguments

py_env

Character. The name, or full path, of the Python environment to be used by rgee.

earthengine_version

Character. The Earth Engine Python API version to install. By default rgee::ee_version().

python_version

Only Windows users. The Python version to be used in this conda environment. If set to NULL, the default Python package will be used. For example, you can specify python_version = "3.6" to request the creation of the conda environment with a copy of Python 3.6.

confirm

Logical. Confirm before restarting R?.

Value

No return value, called for installing non-R dependencies.

See also

Examples

if (FALSE) { # \dontrun{
library(rgee)
# ee_install()
} # }