Interface for manage the Earth Engine Asset

R functions for managing the Earth Engine Asset

ee_manage_create(path_asset, asset_type = "folder", quiet = FALSE)

ee_manage_delete(path_asset, quiet = FALSE)

ee_manage_assetlist(path_asset, quiet = FALSE)

ee_manage_quota()

ee_manage_size(path_asset)

ee_manage_copy(path_asset, final_path, quiet = FALSE)

ee_manage_move(path_asset, final_path, quiet = FALSE)

ee_manage_set_properties(path_asset, properties)

ee_manage_delete_properties(path_asset, property)

ee_manage_assets_access(path_asset,
  acl = getOption("rgee.manage.getAssetAcl"), quiet = FALSE)

ee_manage_task(quiet, cache = TRUE)

ee_manage_cancel_all_running_taks()

Arguments

path_asset
Character. Name of the EE asset e.g. users/user/nameofcollection.

asset_type
Character. The asset type ('folder' or 'imagecollection').

quiet
Logical. Suppress info message.

final_path
Character. Name of the final EE asset to copy or move.

properties
List. Set of parameters to established as a property of an EE object. See details.

property
Character. Name of a specific property to be deleted.

acl
A list describing the asset's Access Control List. See getOption("rgee.manage.getAssetAcl").

cache
Logical. TRUE save the report in the /temp directory.

Examples

library(rgee)
ee_reattach() # reattach ee as a reserved word
ee_Initialize()

# Change google account to be able to reproduce
ee_manage_create('users/aybar1994/rgee')

#> GEE asset: users/aybar1994/rgee created

# 1. List all the elements inside a folder or a ImageCollection
ee_manage_assetlist(path_asset = 'users/aybar1994/rgee')

#> data frame with 0 columns and 0 rows

# 2. Create a Folder or a ImageCollection
ee_manage_create('users/aybar1994/rgee/rgee_folder',asset_type = 'folder')

#> GEE asset: users/aybar1994/rgee/rgee_folder created

ee_manage_create('users/aybar1994/rgee/rgee_ic',asset_type = 'imagecollection')

#> GEE asset: users/aybar1994/rgee/rgee_ic created

ee_manage_assetlist('users/aybar1994/rgee')

#>                                 ID            TYPE
#> 2     users/aybar1994/rgee/rgee_ic ImageCollection
#> 1 users/aybar1994/rgee/rgee_folder          Folder

# 3. Shows your Earth Engine quota

# 4. Estimate the size of a Image, ImageCollection, Table or Folder.
ee_manage_size('users/aybar1994/rgee')

#> Folder :
#>  - Size: 0 B
#>  - # elements: 2

# 5. Move a EE object to another folder
ee_manage_move(path_asset = 'users/aybar1994/rgee/rgee_ic',
               final_path = 'users/aybar1994/rgee/rgee_folder/rgee_ic_moved')

#> Done

ee_manage_assetlist('users/aybar1994/rgee/rgee_folder')

#>                                               ID            TYPE
#> 1 users/aybar1994/rgee/rgee_folder/rgee_ic_moved ImageCollection

# 6. Set properties to an EE object.
ee_manage_set_properties(path_asset = 'users/aybar1994/rgee/rgee_folder/rgee_ic_moved',
                         properties = list(message='hello-world',language = 'R'))
test_ic <- ee$ImageCollection('users/aybar1994/rgee/rgee_folder/rgee_ic_moved')
test_ic$getInfo()

#> $type
#> [1] "ImageCollection"
#> 
#> $bands
#> list()
#> 
#> $version
#> [1] -853244001
#> 
#> $id
#> [1] "users/aybar1994/rgee/rgee_folder/rgee_ic_moved"
#> 
#> $properties
#> $properties$language
#> [1] "R"
#> 
#> $properties$message
#> [1] "hello-world"
#> 
#> 
#> $features
#> list()
#> 

# 7. Delete properties
ee_manage_delete_properties(path_asset = 'users/aybar1994/rgee/rgee_folder/rgee_ic_moved',
                            property = c("message","language"))
test_ic$getInfo()

#> $type
#> [1] "ImageCollection"
#> 
#> $bands
#> list()
#> 
#> $version
#> [1] -851321841
#> 
#> $id
#> [1] "users/aybar1994/rgee/rgee_folder/rgee_ic_moved"
#> 
#> $features
#> list()
#> 

# 8. Share EE objects -- Create a public dataset
ee_manage_assets_access('users/aybar1994/rgee/rgee_folder/rgee_ic_moved')

#> The ACL of users/aybar1994/rgee/rgee_folder/rgee_ic_moved has been changed.

ee$data$getAssetAcl('users/aybar1994/rgee/rgee_folder/rgee_ic_moved')

#> $all_users_can_read
#> [1] TRUE
#> 
#> $readers
#> list()
#> 
#> $writers
#> list()
#> 
#> $owners
#> [1] "aybar1994@gmail.com"
#> 

# 9. Create a report based on all tasks that is running or has finished
ee_manage_task()

#>                          ID     State                       DestinationPath
#> 1  EAUYFIKXLJ5N7ZXRBYU2ST33 COMPLETED         users/aybar1994/testing_01/L7
#> 2  KCVITAPL7DOR4V57SV64VLOP COMPLETED                     myExportImageTask
#> 3  4B5O3XY3B7TASZBPH7CJRMP7 COMPLETED users/aybar1994/testing_01/geomatrix2
#> 4  DPT72IZYKSUOQG6IPGSWSK3D COMPLETED  users/aybar1994/testing_01/geomatrix
#> 5  SBOZRDWVUFKJEHQ2NUKEVS73 COMPLETED  users/aybar1994/testing_01/geomatrix
#> 6  6OWDXE3QG6DIWLQHTSAZRH4Q COMPLETED                     myExportTableTask
#> 7  4T6LDZGKBRBDACRVBENIF4D4 COMPLETED                     myExportImageTask
#> 8  GLJ3O55LOL5QAIBHOCB7MAIT COMPLETED                     myExportImageTask
#> 9  HPUVTUJCL3DW5LQRPUEZWUYM COMPLETED                     myExportImageTask
#> 10 TUI6WRPF7Y2MU3CWXGOQKOWX COMPLETED                     myExportImageTask
#> 11 AUK4AZ6IBHL67BMHNWAA635O COMPLETED                     myExportImageTask
#> 12 P4JCHA6ZNLNXPS3D7LVU6YTS COMPLETED                     myExportImageTask
#> 13 CST635T7CJ27EZCGDM5GT53Z COMPLETED                     myExportImageTask
#> 14 FB5E2WKET33P3SSPKYG3E6SJ    FAILED                     myExportImageTask
#> 15 EN3WEADPR7YRJXCYBSGHBHRG COMPLETED                     myExportImageTask
#> 16 PAAQIJ3D5JDGH6X7CGKL6P6I COMPLETED                     myExportImageTask
#> 17 HDPOLRXBOB4CBT76ESZUU3RR COMPLETED                     myExportImageTask
#> 18 C7O3NWHJSB7JRD5F4747BVQH COMPLETED                     myExportImageTask
#> 19 OWVWJ23MGHOJLSH4TPUPWXTZ COMPLETED                     myExportImageTask
#> 20 MDRDCIRKS5Y7IPFY5XBFYNRV COMPLETED                     myExportImageTask
#> 21 7GBJ4KM4X5KRQESHE7PTM2UU    FAILED                     myExportImageTask
#> 22 A222KL3X5NMVFTQ7QCMW2ZFF COMPLETED                     myExportImageTask
#> 23 HNFU3UH5XXWCWSB4EPYELL6C COMPLETED                     myExportImageTask
#> 24 LDACDU53C2IJKGGCCOV7NFS4 COMPLETED                     myExportImageTask
#> 25 737SW57LSJVRTY5T5EEUTA3N COMPLETED                     myExportImageTask
#> 26 E6Z4725NOEUI2INLYCN6LOCU COMPLETED                     myExportImageTask
#> 27 SDN64B7AW4BRZ3KTCDQTSZKB COMPLETED                     myExportTableTask
#> 28 VMYVXVYFDF5GYRUUMQF6CYNR COMPLETED                     myExportTableTask
#> 29 NQO76YKTBNB3KWJIGZAQVIJ7 COMPLETED                     myExportTableTask
#> 30 P24JICIGELCPIJJUKFULXJCG COMPLETED                     myExportTableTask
#> 31 JV7JQGOVIQNDETBRBIPMPNJZ COMPLETED                     myExportTableTask
#> 32 CUANJ2WSL63G2NJJJ5GJY75G COMPLETED                     myExportTableTask
#> 33 N26IHF4SXSPYQOYS6SBHOI3R COMPLETED                     myExportTableTask
#> 34 QOHMZCVCKOEI3BO46TDGKCFM COMPLETED                     myExportTableTask
#> 35 STSZWIOS32UU2YYPSWLOIMI6 COMPLETED                     myExportTableTask
#> 36 HHZPJ2R5OF6RHF7HQOW3IIXP COMPLETED                     myExportTableTask
#> 37 NPUUN4OUFD5QEXN5I225VXUT COMPLETED                     myExportTableTask
#> 38 XXISPYHPTW62RDSYRPJFNDIH COMPLETED                     myExportTableTask
#> 39 JO54BLLOEALFKYFOP53HVHFO COMPLETED                     myExportTableTask
#> 40 O2KK243TOOKHWIBC2AIEC3H2 COMPLETED                     myExportTableTask
#> 41 N5EOLZI56CHHQPDQ42XFIULH COMPLETED                     myExportImageTask
#> 42 3PE7CFLOWZHNYZBAFA2YZWQL COMPLETED                     myExportImageTask
#> 43 PHE6IXXL2VWXCI3QZJFNV4TM COMPLETED                     myExportImageTask
#> 44 A2LRVKYFBY6LWNIPG7EGY443 COMPLETED                     myExportImageTask
#> 45 UAV4OYEOP342MPP5RQULMHYL COMPLETED                     myExportImageTask
#> 46 QM7ZNIZLWWH5US4QQMU2KOZS COMPLETED                     myExportImageTask
#> 47 PVIK3OSKFYRW6LWR4PWAGMU6 COMPLETED                     myExportImageTask
#>               Type               Start DeltaToCreate(s) DeltaToCompletedTask(s)
#> 1           INGEST 2019-09-12 23:10:54 11.7979998588562        26.5090000629425
#> 2     EXPORT_IMAGE 2019-09-12 18:46:45 10.3159999847412        37.4019999504089
#> 3           INGEST 2019-09-12 18:03:40 12.9820001125336        26.8489999771118
#> 4           INGEST 2019-09-12 17:58:33 9.91500020027161        23.0909998416901
#> 5           INGEST 2019-09-12 17:53:50 6.79500007629395        21.1340000629425
#> 6  EXPORT_FEATURES 2019-09-09 18:12:05 4.92899990081787        2.20499992370605
#> 7     EXPORT_IMAGE 2019-09-09 18:09:08  4.5550000667572        9.78799986839294
#> 8     EXPORT_IMAGE 2019-09-09 17:55:14 11.9470000267029        9.83999991416931
#> 9     EXPORT_IMAGE 2019-09-08 03:31:46 11.3650000095367        16.2409999370575
#> 10    EXPORT_IMAGE 2019-09-08 03:30:23  3.5310001373291        17.7999999523163
#> 11    EXPORT_IMAGE 2019-09-08 03:18:50 3.54299998283386        15.4100000858307
#> 12    EXPORT_IMAGE 2019-09-08 03:11:03 5.36600017547607        16.7009999752045
#> 13    EXPORT_IMAGE 2019-09-08 03:10:27 5.14199995994568        17.6380000114441
#> 14    EXPORT_IMAGE 2019-09-08 03:08:27 7.04299998283386        1.41100001335144
#> 15    EXPORT_IMAGE 2019-09-08 03:05:32 5.39999985694885        14.4219999313354
#> 16    EXPORT_IMAGE 2019-09-08 03:05:03 4.68700003623962        14.8159999847412
#> 17    EXPORT_IMAGE 2019-09-08 03:04:27 6.07500004768372        20.2909998893738
#> 18    EXPORT_IMAGE 2019-09-08 03:00:37 5.46000003814697        16.9300000667572
#> 19    EXPORT_IMAGE 2019-09-08 02:59:29 9.80199980735779        51.6560001373291
#> 20    EXPORT_IMAGE 2019-09-08 02:58:53 3.35199999809265        14.0789999961853
#> 21    EXPORT_IMAGE 2019-09-08 02:57:17  13.268000125885       0.601999998092651
#> 22    EXPORT_IMAGE 2019-09-08 02:51:43 3.43099999427795        14.4810001850128
#> 23    EXPORT_IMAGE 2019-09-08 02:51:13 4.04900002479553         15.194000005722
#> 24    EXPORT_IMAGE 2019-09-08 02:50:38 7.54500007629395        15.3529999256134
#> 25    EXPORT_IMAGE 2019-09-08 02:49:22 7.19599986076355        14.3140001296997
#> 26    EXPORT_IMAGE 2019-09-08 02:35:50 15.3309998512268        376.728000164032
#> 27 EXPORT_FEATURES 2019-09-07 06:12:40 5.85099983215332        1.69800019264221
#> 28 EXPORT_FEATURES 2019-09-07 06:12:26  7.8970000743866        2.32999992370605
#> 29 EXPORT_FEATURES 2019-09-07 06:12:13 9.83699989318848        1.64900016784668
#> 30 EXPORT_FEATURES 2019-09-07 06:11:53 10.1949999332428        1.66700005531311
#> 31 EXPORT_FEATURES 2019-09-07 06:11:34 12.7540001869202         1.6399998664856
#> 32 EXPORT_FEATURES 2019-09-07 06:11:14 5.73200011253357        6.05299997329712
#> 33 EXPORT_FEATURES 2019-09-07 06:11:01 7.29299998283386        1.51300001144409
#> 34 EXPORT_FEATURES 2019-09-07 06:10:51 3.46000003814697         2.0699999332428
#> 35 EXPORT_FEATURES 2019-09-07 06:10:31 10.1449999809265        1.99900007247925
#> 36 EXPORT_FEATURES 2019-09-07 06:10:17 9.15100002288818        2.16499996185303
#> 37 EXPORT_FEATURES 2019-09-07 06:10:07 3.53800010681152        2.17799997329712
#> 38 EXPORT_FEATURES 2019-09-07 06:09:53 8.69400000572205        2.35899996757507
#> 39 EXPORT_FEATURES 2019-09-07 06:09:27 6.53999996185303         11.779000043869
#> 40 EXPORT_FEATURES 2019-09-07 06:09:12 7.64399981498718        2.40200018882751
#> 41    EXPORT_IMAGE 2019-09-07 06:08:36 5.07200002670288        28.0920000076294
#> 42    EXPORT_IMAGE 2019-09-07 06:07:55 10.3000001907349        28.4229998588562
#> 43    EXPORT_IMAGE 2019-09-07 06:07:30 9.82100009918213        9.22399997711182
#> 44    EXPORT_IMAGE 2019-09-07 06:06:48 7.98099994659424        27.6900000572205
#> 45    EXPORT_IMAGE 2019-09-07 06:06:17 5.75399994850159        21.8450000286102
#> 46    EXPORT_IMAGE 2019-09-07 06:05:51  5.9539999961853        9.65599989891052
#> 47    EXPORT_IMAGE 2019-09-07 05:00:06 8.23399996757507        11.5150001049042
#>                                                                                                          ErrorMessage
#> 1                                                                                                                NULL
#> 2                                                                                                                NULL
#> 3                                                                                                                NULL
#> 4                                                                                                                NULL
#> 5                                                                                                                NULL
#> 6                                                                                                                NULL
#> 7                                                                                                                NULL
#> 8                                                                                                                NULL
#> 9                                                                                                                NULL
#> 10                                                                                                               NULL
#> 11                                                                                                               NULL
#> 12                                                                                                               NULL
#> 13                                                                                                               NULL
#> 14                                                                    Image.reproject: Parameter 'image' is required.
#> 15                                                                                                               NULL
#> 16                                                                                                               NULL
#> 17                                                                                                               NULL
#> 18                                                                                                               NULL
#> 19                                                                                                               NULL
#> 20                                                                                                               NULL
#> 21 Collection.first: Error in map(ID=LC08_003069_20160730):\nImage.bitwiseAnd: Bitwise operands must be integer only.
#> 22                                                                                                               NULL
#> 23                                                                                                               NULL
#> 24                                                                                                               NULL
#> 25                                                                                                               NULL
#> 26                                                                                                               NULL
#> 27                                                                                                               NULL
#> 28                                                                                                               NULL
#> 29                                                                                                               NULL
#> 30                                                                                                               NULL
#> 31                                                                                                               NULL
#> 32                                                                                                               NULL
#> 33                                                                                                               NULL
#> 34                                                                                                               NULL
#> 35                                                                                                               NULL
#> 36                                                                                                               NULL
#> 37                                                                                                               NULL
#> 38                                                                                                               NULL
#> 39                                                                                                               NULL
#> 40                                                                                                               NULL
#> 41                                                                                                               NULL
#> 42                                                                                                               NULL
#> 43                                                                                                               NULL
#> 44                                                                                                               NULL
#> 45                                                                                                               NULL
#> 46                                                                                                               NULL
#> 47                                                                                                               NULL

# 10. Cancel all the running task
ee_manage_cancel_all_running_taks()

# 11. Delete EE objects or folders
ee_manage_delete('users/aybar1994/rgee/')

#> EE object deleted: users/aybar1994/rgee/rgee_folder/rgee_ic_moved 
#> EE object deleted: users/aybar1994/rgee/rgee_folder 
#> EE object deleted: users/aybar1994/rgee