Interface for manage the Earth Engine Asset

R functions for managing the Earth Engine Asset

ee_manage_create(path_asset, asset_type = "Folder", quiet = FALSE)

ee_manage_delete(path_asset, quiet = FALSE)

ee_manage_assetlist(path_asset, quiet = FALSE)

ee_manage_quota()

ee_manage_copy(path_asset, final_path, quiet = FALSE)

ee_manage_move(path_asset, final_path, quiet = FALSE)

ee_manage_set_properties(path_asset, add_properties)

ee_manage_delete_properties(path_asset, del_properties = "ee_rmv_all_prp")

ee_manage_asset_access(
  path_asset,
  editor = NULL,
  viewer = NULL,
  all_users_can_read = TRUE,
  quiet = FALSE
)

ee_manage_task(cache = FALSE)

ee_manage_cancel_all_running_task()

ee_manage_asset_size(path_asset)

Arguments

path_asset
Character. Name of the EE asset (Table, Image, Folder or ImageCollection).

asset_type
Character. The asset type to create ('Folder' or 'ImageCollection').

quiet
Logical. Suppress info message.

final_path
Character. Output filename (e.g users/datacolecfbf/ic_moved)

add_properties
List. Set of parameters to established as a property of an EE object. See details.

del_properties
Character. Names of properties to be deleted. See details.

editor
Character vector. Define editor users in the IAM Policy.

viewer
Character vector. Define viewer users in the IAM Policy.

all_users_can_read
Logical. All users can see the asset element.

cache
Logical. If TRUE, the task report will be saved in the /temp directory and used when the function .

Details

If the argument del_properties is 'ee_rmv_all_prp', ee_manage_delete_properties will delete all the properties.

Examples

if (FALSE) {
library(rgee)
ee_Initialize()
ee_user_info()

# Change google account to be able to reproduce

# 1. Create a folder or Image Collection
# Change path asset according to your specific user
ee_manage_create("users/datacolecfbf/rgee")

# 1. List all the elements inside a folder or a ImageCollection
ee_manage_assetlist(path_asset = "users/datacolecfbf/rgee")

# 2. Create a Folder or a ImageCollection
ee_manage_create(
  path_asset = "users/datacolecfbf/rgee/rgee_folder",
  asset_type = "Folder"
)

ee_manage_create(
  path_asset = "users/datacolecfbf/rgee/rgee_ic",
  asset_type = "ImageCollection"
)

ee_manage_assetlist(path_asset = "users/datacolecfbf/rgee")

# 3. Shows Earth Engine quota
ee_manage_quota()

# 4. Move an EE object to another folder
ee_manage_move(
  path_asset = "users/datacolecfbf/rgee/rgee_ic",
  final_path = "users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved"
)

ee_manage_assetlist(path_asset = "users/datacolecfbf/rgee/rgee_folder")

# 5. Set properties to an EE object.
ee_manage_set_properties(
  path_asset = "users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved",
  add_properties = list(message = "hello-world", language = "R")
)

ic_id <- "users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved"
test_ic <- ee$ImageCollection(ic_id)
test_ic$getInfo()

# 6. Delete properties
ee_manage_delete_properties(
  path_asset = "users/datacolecfbf/rgee/rgee_folder/rgee_ic_moved",
  del_properties = c("message", "language")
)
test_ic$getInfo()

# 7. Create a report based on all the tasks
# that are running or have already been completed.
ee_manage_task()

# 8. Cancel all the running task
ee_manage_cancel_all_running_task()

# 9. Delete EE objects or folders
ee_manage_delete("users/datacolecfbf/rgee/")
}