Adds a given EE object to mapview as a layer.

Display a given EE object (Geometry, Image, Feature, FeatureCollection or ImageCollection) using the mapview style.

ee_map(eeobject, vizparams, center, zoom_start = 8, obj_name = "map",
  max_nimage = 10, ...)

# S3 method for default
ee_map(eeobject, ...)

# S3 method for ee.geometry.Geometry
ee_map(eeobject, vizparams, center,
  zoom_start = 8, obj_name = "map", ...)

# S3 method for ee.feature.Feature
ee_map(eeobject, vizparams, center,
  zoom_start = 8, obj_name = "map", ...)

# S3 method for ee.featurecollection.FeatureCollection
ee_map(eeobject,
  vizparams, center, zoom_start = 8, obj_name = "map", ...)

# S3 method for ee.image.Image
ee_map(eeobject, vizparams, center,
  zoom_start = 2, obj_name = "map", ...)

# S3 method for ee.imagecollection.ImageCollection
ee_map(eeobject, vizparams,
  center, zoom_start = 2, obj_name = "map", max_nimage = 10, ...)

Arguments

eeobject
EE object.

vizparams
list; visualization parameters. See Details.

center
character vector; the longitude and latitude of the map center.

zoom_start
zoom level.

obj_name
character vector; name of the map(or maps).

max_nimage
Max number of Image to display.

...
ignored

Details

ee_map takes advantage of the getMapId function for generating a provisial WMS service supported by Google Earth Engine. To achieve desirable visualization effects, it is depend on the type of EE object. For neither Image or ImageCollection, you can provide visualization parameters to ee_map by the parameter vizparams. The parameters available are:

Parameter

Description

Type

bands		Comma-delimited list of three band names to be mapped to RGB

list		min

Value(s) to map to 0	number or list of three numbers, one for each band

max

Value(s) to map to 255	number or list of three numbers, one for each band

gain		Value(s) by which to multiply each pixel value

number or list of three numbers, one for each band		bias

Value(s) to add to each DN	number or list of three numbers, one for each band

gamma

Gamma correction factor(s)	number or list of three numbers, one for each band

palette		List of CSS-style color strings (single-band images only)

comma-separated list of hex strings		opacity

The opacity of the layer (0.0 is fully transparent and 1.0 is fully opaque)	number

format

Either "jpg" or "png"	string

If you add a EE Image or ImageCollection object to the map without any additional parameters, by default ee_map assigns the first three bands to red, green and blue, respectively. The default stretch is based on the mixmax range , which may or may not be suitable. In the case of Geometry, Feature or FeatureCollection objects vizparams allow the next parameters:

color: A hex string in the format RRGGBB specifying the color to use for drawing the features. By default 000000.

pointRadius: The radius of the point markers. By default 3.

strokeWidth: The width of lines and polygon borders. By default 3.

Examples

# NOT RUN {
library(rgee)
ee_initialize()

# Case: Geometry*
geom <- ee$Geometry$Point(list(-73.53522, -15.75453))
m1 <- ee_map(geom,list(pointRadius=10,color="FF0000"), obj_name = "Geometry-Arequipa")
m1

# Case: Feature
eeobject_fc <- ee_FeatureCollection("users/csaybar/DLdemos/train_set")$first()
m2 <- ee_map(ee$Feature(eeobject_fc), obj_name = "Feature-Arequipa")
m2 + m1

# Case: FeatureCollection
eeobject_fc <- ee_FeatureCollection("users/csaybar/DLdemos/train_set")
m3 <- ee_map(eeobject_fc,obj_name = 'FeatureCollection')
m3+m1

# Case: Image
image <- ee$Image("LANDSAT/LC08/C01/T1/LC08_044034_20140318")
m4 <- ee_map(image,list(bands = c('B4','B3','B2'), max=10000),obj_name = "Image")
m4+m1

# Case: ImageCollection
collection <- ee_ImageCollection("LANDSAT/LC08/C01/T1_TOA")$
  filter(ee_Filter()$eq("WRS_PATH", 44))$
  filter(ee_Filter()$eq("WRS_ROW", 34))$
  filterDate("2014-01-01", "2015-01-01")$
  sort("CLOUD_COVER")

m5 <- ee_map(eeobject = collection,
             vizparams = list(bands = c('B4','B3','B2'), max=1),
             #obj_name = "ImageCollection",
             obj_name = c("Scene_2019","Scene_2016","Scene_2011"),
             max_nimage = 3)

m5 + m1 + m2 + m3 + m4
# }