Print and return metadata about Spatial Earth Engine Objects

Print and return metadata about Spatial Earth Engine Objects

ee_print(eeobject, clean = FALSE, max_display = 0)

# S3 method for ee.geometry.Geometry
ee_print(eeobject, clean = FALSE,
  max_display = 0)

# S3 method for ee.feature.Feature
ee_print(eeobject, clean = FALSE,
  max_display = 0)

# S3 method for ee.featurecollection.FeatureCollection
ee_print(eeobject,
  clean = FALSE, max_display = 0)

# S3 method for ee.image.Image
ee_print(eeobject, clean = FALSE,
  max_display = 0)

# S3 method for ee.imagecollection.ImageCollection
ee_print(eeobject,
  clean = FALSE, max_display = 0)

Arguments

eeobject
Earth Engine Object. Available for: Geometry, Feature, FeatureCollection, Image or ImageCollection.

clean
Logical; Whether is TRUE cache will cleaned, see Description.

max_display
Set the max number of properties to display.

Details

For avoiding computing time extremely large and "EEException" related with memory consuming (See this Earth Engine Guide) just the first element of: ee$Image, ee$ImageCollection, ee$Feature and ee$FeatureCollection will be used to generated Earth Engine object properties. By default ee_print create a list with all the print parameters. It is due to avoid repeated queries to the Earth Engine Server.

Examples

library(rgee)
ee_Initialize()
eeobject <- ee$ImageCollection("LANDSAT/LC08/C01/T1_TOA")$
  filter(ee$Filter()$eq("WRS_PATH", 44))$
  filter(ee$Filter()$eq("WRS_ROW", 34))$
  filterDate("2014-03-01", "2014-08-01")
ee_print(eeobject, max_display = 118)

#> Class                          : ImageCollection 
#> nImages                        : 9 
#> nBands                         : 12 
#> Bands names                    : B1 B2 B3 B4 B5 B6 B7 B8 B9 B10 B11 BQA 
#> EPSG (SRID)                    : 32610 
#> proj4string                    : +proj=utm +zone=10 +datum=WGS84 +units=m +no_defs 
#> Geotransform                   : [30 0 460785 0 -30 4264215] 
#> Scale (m)                      : 30 
#> Dimensions                     : nrow:7661  ncol:7801 
#> nPixels by image               : 59'763'461 
#> Data type                      : FLOAT 
#> Image Properties               : 118 
#>                                |           THERMAL_LINES            |         EARTH_SUN_DISTANCE         |
#>                                |          LANDSAT_SCENE_ID          |         SCENE_CENTER_TIME          |
#>                                |            SUN_AZIMUTH             |       REFLECTANCE_ADD_BAND_6       |
#>                                |         PANCHROMATIC_LINES         |         system:time_start          |
#>                                |             REQUEST_ID             |           BPF_NAME_TIRS            |
#>                                |          ELEVATION_SOURCE          |       GEOMETRIC_RMSE_VERIFY        |
#>                                |        RADIANCE_MULT_BAND_7        |        RADIANCE_MULT_BAND_6        |
#>                                |        RADIANCE_MULT_BAND_5        |        RADIANCE_MULT_BAND_4        |
#>                                |        RADIANCE_MULT_BAND_3        |        RADIANCE_MULT_BAND_2        |
#>                                |        RADIANCE_MULT_BAND_1        |        GEOMETRIC_RMSE_MODEL        |
#>                                |               DATUM                |        RADIANCE_MULT_BAND_9        |
#>                                |        RADIANCE_MULT_BAND_8        |           RLUT_FILE_NAME           |
#>                                |       RADIANCE_MULT_BAND_11        |       RADIANCE_MULT_BAND_10        |
#>                                |        PANCHROMATIC_SAMPLES        |   GROUND_CONTROL_POINTS_VERSION    |
#>                                |           TRUNCATION_OLI           |             ELLIPSOID              |
#>                                |    PROCESSING_SOFTWARE_VERSION     |          TARGET_WRS_PATH           |
#>                                |       REFLECTANCE_ADD_BAND_8       |       REFLECTANCE_ADD_BAND_7       |
#>                                |           SUN_ELEVATION            |        COLLECTION_CATEGORY         |
#>                                | TIRS_STRAY_LIGHT_CORRECTION_SOURCE |        RADIANCE_ADD_BAND_11        |
#>                                |       REFLECTANCE_ADD_BAND_5       |         IMAGE_QUALITY_OLI          |
#>                                |    GROUND_CONTROL_POINTS_VERIFY    |       REFLECTANCE_ADD_BAND_4       |
#>                                |       REFLECTANCE_ADD_BAND_3       |         system:asset_size          |
#>                                |             FILE_DATE              |        RADIANCE_ADD_BAND_4         |
#>                                |         RESAMPLING_OPTION          |       REFLECTANCE_ADD_BAND_1       |
#>                                |              WRS_PATH              |           DATE_ACQUIRED            |
#>                                |            system:index            |         SATURATION_BAND_1          |
#>                                |           NADIR_OFFNADIR           |             DATA_TYPE              |
#>                                |         REFLECTIVE_SAMPLES         |     GRID_CELL_SIZE_REFLECTIVE      |
#>                                |    GRID_CELL_SIZE_PANCHROMATIC     |         SATURATION_BAND_7          |
#>                                |           TARGET_WRS_ROW           |      TIRS_SSM_POSITION_STATUS      |
#>                                |          system:footprint          |    GROUND_CONTROL_POINTS_MODEL     |
#>                                |          REFLECTIVE_LINES          |           TIRS_SSM_MODEL           |
#>                                |            CLOUD_COVER             |         IMAGE_QUALITY_TIRS         |
#>                                |           SPACECRAFT_ID            |             STATION_ID             |
#>                                |        RADIANCE_ADD_BAND_9         |              CPF_NAME              |
#>                                |       REFLECTANCE_ADD_BAND_9       |     google:registration_ratio      |
#>                                |        RADIANCE_ADD_BAND_1         |            BPF_NAME_OLI            |
#>                                |        RADIANCE_ADD_BAND_3         |        RADIANCE_ADD_BAND_2         |
#>                                |        RADIANCE_ADD_BAND_5         |       REFLECTANCE_ADD_BAND_2       |
#>                                |        RADIANCE_ADD_BAND_7         |        RADIANCE_ADD_BAND_6         |
#>                                |      REFLECTANCE_MULT_BAND_9       |      REFLECTANCE_MULT_BAND_8       |
#>                                |             SENSOR_ID              |         LANDSAT_PRODUCT_ID         |
#>                                |      REFLECTANCE_MULT_BAND_1       |      REFLECTANCE_MULT_BAND_3       |
#>                                |      REFLECTANCE_MULT_BAND_2       |      REFLECTANCE_MULT_BAND_5       |
#>                                |      REFLECTANCE_MULT_BAND_4       |      REFLECTANCE_MULT_BAND_7       |
#>                                |      REFLECTANCE_MULT_BAND_6       |              WRS_ROW               |
#>                                |        RADIANCE_ADD_BAND_8         |        K1_CONSTANT_BAND_10         |
#>                                |        K1_CONSTANT_BAND_11         |          CLOUD_COVER_LAND          |
#>                                |     google:registration_count      |       GEOMETRIC_RMSE_MODEL_X       |
#>                                |       GEOMETRIC_RMSE_MODEL_Y       |          THERMAL_SAMPLES           |
#>                                |            ORIENTATION             |              UTM_ZONE              |
#>                                |         SATURATION_BAND_8          |         SATURATION_BAND_9          |
#>                                |           MAP_PROJECTION           |        RADIANCE_ADD_BAND_10        |
#>                                |         SATURATION_BAND_2          |         SATURATION_BAND_3          |
#>                                |       GRID_CELL_SIZE_THERMAL       |         SATURATION_BAND_6          |
#>                                |         COLLECTION_NUMBER          |         SATURATION_BAND_4          |
#>                                |         SATURATION_BAND_5          |    google:registration_offset_y    |
#>                                |    google:registration_offset_x    |             ROLL_ANGLE             |
#>                                |        K2_CONSTANT_BAND_11         |        K2_CONSTANT_BAND_10         |
#> ImageCollection Properties     : 39 
#>                                |         provider_url         | system:visualization_0_bands |
#>                                | system:visualization_2_bands |            period            |
#>                                |     visualization_2_max      |  system:visualization_1_max  |
#>                                | system:visualization_1_bands |     visualization_0_max      |
#>                                |     visualization_2_name     |     visualization_1_max      |
#>                                |     visualization_0_gain     |     visualization_0_min      |
#>                                |     visualization_1_name     |  system:visualization_0_max  |
#>                                |            thumb             |            title             |
#>                                |          type_name           |  system:visualization_2_max  |
#>                                |     visualization_2_min      |    visualization_0_bands     |
#>                                |           provider           | system:visualization_1_name  |
#>                                |  system:visualization_0_min  |  system:visualization_2_min  |
#>                                |         description          |         source_tags          |
#>                                | system:visualization_2_name  | system:visualization_0_name  |
#>                                |          date_range          |     visualization_0_name     |
#>                                |     visualization_2_gain     |     visualization_1_gain     |
#>                                |    visualization_1_bands     |             tags             |
#>                                |         product_tags         |            sample            |
#>                                |    visualization_2_bands     |     visualization_1_min      |
#>                                |  system:visualization_1_min  |