Print and return metadata about Spatial Earth Engine Objects

Print and return metadata about Spatial Earth Engine Objects. ee_print can retrieve information about the number of images or features, number of bands or geometries, number of pixels, geotransform, data type, properties and object size.

ee_print(eeobject, ...)

# S3 method for ee.geometry.Geometry
ee_print(eeobject, ..., clean = FALSE, quiet = FALSE)

# S3 method for ee.feature.Feature
ee_print(eeobject, ..., clean = FALSE, quiet = FALSE)

# S3 method for ee.featurecollection.FeatureCollection
ee_print(eeobject, ..., f_index = 0, clean = FALSE, quiet = FALSE)

# S3 method for ee.image.Image
ee_print(
  eeobject,
  ...,
  img_band,
  time_end = FALSE,
  compression_ratio = 20,
  clean = FALSE,
  quiet = FALSE
)

# S3 method for ee.imagecollection.ImageCollection
ee_print(
  eeobject,
  ...,
  time_end = FALSE,
  img_index = 0,
  img_band,
  compression_ratio = 20,
  clean = FALSE,
  quiet = FALSE
)

Arguments

eeobject
Earth Engine Object. Available for: Geometry, Feature, FeatureCollection, Image or ImageCollection.

...
ignored

clean
Logical. If TRUE, the cache will be cleaned.

quiet
Logical. Suppress info message

f_index
Numeric. Index of the ee$FeatureCollection to fetch. Relevant just for ee$FeatureCollection objects.

img_band
Character. Band name of the ee$Image to fetch. Relevant just for ee$ImageCollection and ee$Image objects.

time_end
Logical. If TRUE, the system:time_end property in ee$Image is also returned. See rgee::ee_get_date_img for details.

compression_ratio
Numeric. Measurement of the relative reduction in size of data representation produced by a data compression algorithm (ignored if eeobject is not a Image or ImageCollection). By default is 20.

img_index
Numeric. Index of the ee$ImageCollection to fetch. Relevant just for ee$ImageCollection objects.

Value

A list with the metadata of the Earth Engine object.

See also

Examples

if (FALSE) {
library(rgee)
ee_Initialize()

# Geometry
geom <- ee$Geometry$Rectangle(-10,-10,10,10)
Map$addLayer(geom)
ee_print(geom)

# Feature
feature <- ee$Feature(geom, list(rgee = "ee_print", data = TRUE))
ee_print(feature)

# FeatureCollection
featurecollection <- ee$FeatureCollection(feature)
ee_print(featurecollection)

# Image
srtm <- ee$Image("CGIAR/SRTM90_V4")
ee_print(srtm)

srtm_clip <- ee$Image("CGIAR/SRTM90_V4")$clip(geom)
srtm_metadata <- ee_print(srtm_clip)
srtm_metadata$img_bands_names

# ImageCollection
object <- ee$ImageCollection("LANDSAT/LC08/C01/T1_TOA")$
  filter(ee$Filter()$eq("WRS_PATH", 44))$
  filter(ee$Filter()$eq("WRS_ROW", 34))$
  filterDate("2014-03-01", "2014-08-01")$
  aside(ee_print)
}