Print and return metadata about Spatial Earth Engine Objects

Print and return metadata about Spatial Earth Engine Objects. By default ee_print will create a table printing all the available metadata.

ee_print(eeobject, clean = FALSE, max_display = 0)

# S3 method for ee.geometry.Geometry
ee_print(eeobject, clean = FALSE,
  max_display = 0)

# S3 method for ee.feature.Feature
ee_print(eeobject, clean = FALSE,
  max_display = 0)

# S3 method for ee.featurecollection.FeatureCollection
ee_print(eeobject,
  clean = FALSE, max_display = 0)

# S3 method for ee.image.Image
ee_print(eeobject, clean = FALSE,
  max_display = 0)

# S3 method for ee.imagecollection.ImageCollection
ee_print(eeobject,
  clean = FALSE, max_display = 0)

Arguments

eeobject
Earth Engine Object. Available for: Geometry, Feature, FeatureCollection, Image or ImageCollection.

clean
Logical. If TRUE, the cache will be cleaned.

max_display
Set the max number of properties to display.

Examples

library(rgee)

ee_reattach() # reattach ee as a reserved word
ee_Initialize()

#> ── rgee 0.3.3 ───────────────── earthengine-api 0.1.216 ──

#>  ✔ email: not_defined
#>  ✔ Initializing Google Earth Engine: ✔ Initializing Google Earth Engine:  DONE!
#>  ✔ Earth Engine user: users/datacolecfbf 

#> ──────────────────────────────────────────────────────────

eeobject <- ee$ImageCollection("LANDSAT/LC08/C01/T1_TOA")$
  filter(ee$Filter()$eq("WRS_PATH", 44))$
  filter(ee$Filter()$eq("WRS_ROW", 34))$
  filterDate("2014-03-01", "2014-08-01")
ee_print(eeobject, max_display = 0)

#> ────────────────────────────────────────────────────────── 
#> Class                      : ImageCollection 
#> nImages                    : 9 
#> nBands*                    : 12 
#> Bands names*               : B1 B2 B3 B4 B5 B6 B7 B8 B9 B10 B11 BQA 
#> EPSG (SRID)*               : 32610 
#> proj4string*               : +proj=utm +zone=10 +datum=WGS84 +units=m +no_defs 
#> Geotransform*              : [30 0 460785 0 -30 4264215] 
#> Scale (m)*                 : 30 
#> Dimensions*                : 7661 x 7801 (nrow x ncol): 
#> nPixels by image*          : 59'763'461 
#> Data type*                 : FLOAT 
#> Image Properties*          : 118 
#> ImageCollection Properties : 39 
#> ────────────────────────────────────────────────────────── 
#> * Note: Calculated from the first element (ee.Image)