Pass an Earth Engine date object to R

Pass an Earth Engine date object to R

eedate_to_rdate(ee_date, timestamp = FALSE)

Arguments

ee_date
EE date object (ee$Date)

timestamp
Logical. If TRUE, return the date in milliseconds from the Unix Epoch (1970-01-01 00:00:00 UTC) otherwise return the date as a POSIXct object. By default FALSE.

Value

eedate_to_rdate will return a numeric timestamp or an POSIXct object depending on the timestamp argument.

Details

eedate_to_rdate is essential to avoid potential errors which could appear when users call to retrieve the date. Currently, R integer only support 32 bit signed, such integers can only count up to about 2 billion. This range is extremely insufficient to deal with Google Earth Engine date, which is represent by timestamp in milliseconds since the UNIX epoch. eedate_to_rdate use Python as a backend to obtain the date and convert it in float before to export to R.

Examples

library(rgee)
ee_reattach()

#> Error in py_module_import(module, convert = convert): ImportError: No module named ee

ee_Initialize()

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#> 

eeDate <- ee$Date$fromYMD(2010,1,1)

#> Error: Python module ee was not found.
#> 
#> Detected Python configuration:
#> 
#> python:         /usr/bin/python
#> libpython:      /System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/config/libpython2.7.dylib
#> pythonhome:     /System/Library/Frameworks/Python.framework/Versions/2.7:/System/Library/Frameworks/Python.framework/Versions/2.7
#> version:        2.7.16 (default, Feb 29 2020, 01:55:37)  [GCC 4.2.1 Compatible Apple LLVM 11.0.3 (clang-1103.0.29.20) (-macos10.15-objc-
#> numpy:          /System/Library/Frameworks/Python.framework/Versions/2.7/Extras/lib/python/numpy
#> numpy_version:  1.8.0
#> ee:             [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/bin/python3
#>  /usr/bin/python
#>  /usr/local/bin/python
#> 

eedate_to_rdate(eeDate,timestamp = TRUE) # good

#> Error in py_module_import(module, convert = convert): ImportError: No module named ee
#> 
#> Detailed traceback: 
#>   File "/Users/runner/runners/2.169.1/work/_temp/Library/rgee/python/ee_utils.py", line 8, in <module>
#>     import ee

eeDate$getInfo()$value # bad

#> Error in eval(expr, envir, enclos): object 'eeDate' not found