rgee: An R package for interacting with Google Earth Engine

Google Earth Engine (Gorelick et al., 2017) is a cloud computing platform designed for planetary-scale environmental data analysis that only can be accessed via the Earth Engine code editor, third-party web apps, and the JavaScript and Python client libraries. rgee is a non-official client library for R that uses reticulate to wrap the Earth Engine Python API and provide R users with a familiar interface, rapid development features, and flexibility to analyze data using open-source, R third-party packages.

Details

The package implements and supports:

Earth Engine Module

Install or set all rgee dependencies

Check non-R dependencies

Clean non-R dependencies

Session management

Transform an R Date to an EE Date or vice versa

Create Interactive visualization Maps

Image download

Vector download

Generic download

Assets management

Upload raster

Upload vector

Upload generic

Extract values

Helper functions

Utils functions

I. Earth Engine Module

Interface to main Earth Engine module. Provides access to top level classes and functions as well as sub-modules (e.g. ee$Image, ee$FeatureCollection$first, etc.).

—————————	—————————————————————————————————

ee	Main Earth Engine module.

—————————	—————————————————————————————————

II. Install or set non-R rgee dependencies

—————————	—————————————————————————————————

ee_install	Create an isolated Python virtual environment with all rgee dependencies.

ee_install_set_pyenv	Configure which version of Python to use with rgee.

ee_install_upgrade	Upgrade the Earth Engine Python API.

—————————	—————————————————————————————————

III. Check non-R dependencies

—————————	—————————————————————————————————

ee_check	Check all non-R dependencies.

ee_check_python	Check Python environment.

ee_check_credentials	Check Google credentials.

ee_check_python_packages	Check Python packages: earthengine-api and numpy.

—————————	—————————————————————————————————

IV. Clean container, credentials, or rgee system variables

—————————	—————————————————————————————————

ee_clean_container	Delete files from either a Folder or a Bucket.

ee_clean_pyenv	Remove rgee system variables from .Renviron.

—————————	—————————————————————————————————

V. Session management

—————————	—————————————————————————————————

ee_Initialize	Authenticate and Initialize Earth Engine.

ee_version	Earth Engine API version.

ee_user_info	Display the credentials and general info of the initialized user.

ee_users	Display the credentials of all users as a table.

ee_get_assethome	Get the Asset home name.

ee_get_earthengine_path	Get the path where the credentials are stored.

—————————	—————————————————————————————————

VII. Transform an R Date to an EE Date or vice versa

—————————	—————————————————————————————————

eedate_to_rdate	Pass an Earth Engine date object to R.

rdate_to_eedate	Pass an R date object to Earth Engine.

ee_get_date_img	Get the date of a EE Image.

ee_get_date_ic	Get the date of a EE ImageCollection.

—————————	—————————————————————————————————

VIII. Visualization Map

—————————	—————————————————————————————————

Map	R6 object (Map) to display Earth Engine (EE) spatial objects.

R6Map	R6 class to display Earth Engine (EE) spatial objects.

—————————	—————————————————————————————————

IX. Image download

—————————	—————————————————————————————————

ee_as_raster	Convert an Earth Engine (EE) image in a raster object.

ee_as_stars	Convert an Earth Engine (EE) image in a stars object.

ee_as_thumbnail	Create an R spatial gridded object from an EE thumbnail image.

ee_image_to_asset	Creates a task to export an EE Image to their EE Assets.

ee_image_to_drive	Creates a task to export an EE Image to Drive.

ee_image_to_gcs	Creates a task to export an EE Image to Google Cloud Storage.

ee_image_info	Approximate size of an EE Image object.

ee_imagecollection_to_local	Save an EE ImageCollection in their local system.

—————————	—————————————————————————————————

X. Vector download

—————————	—————————————————————————————————

ee_as_sf	Convert an Earth Engine table in an sf object.

ee_table_to_asset	Creates a task to export a FeatureCollection to an EE table asset.

ee_table_to_drive	Creates a task to export a FeatureCollection to Google Drive.

ee_table_to_gcs	Creates a task to export a FeatureCollection to Google Cloud Storage.

—————————	—————————————————————————————————

XI. Generic download

—————————	—————————————————————————————————

ee_drive_to_local	Move results from Google Drive to a local directory.

ee_gcs_to_local	Move results from Google Cloud Storage to a local directory.

—————————	—————————————————————————————————

XII. Assets management

—————————	—————————————————————————————————

ee_manage-tools	Interface to manage the Earth Engine Asset.

—————————	—————————————————————————————————

XIII. Upload raster

—————————	—————————————————————————————————

stars_as_ee	Convert a stars or stars-proxy object into an EE Image object.

raster_as_ee	Convert a Raster* object into an EE Image object.

gcs_to_ee_image	Move a GeoTIFF image from GCS to their EE assets.

—————————	—————————————————————————————————

XIV. Upload vector

—————————	—————————————————————————————————

gcs_to_ee_table	Move a zipped shapefile from GCS to their EE Assets.

sf_as_ee	Convert an sf object to an EE object.

—————————	—————————————————————————————————

XV. Upload generic

—————————	—————————————————————————————————

local_to_gcs	Upload local files to Google Cloud Storage.

—————————	—————————————————————————————————

XVI. Extract values

—————————	—————————————————————————————————

ee_extract	Extract values from EE Images or ImageCollections objects.

—————————	—————————————————————————————————

XVII. Helper functions

—————————	—————————————————————————————————

ee_help	Documentation for Earth Engine Objects.

ee_print	Print and return metadata about Spatial Earth Engine Objects.

ee_monitoring	Monitoring Earth Engine task progress.

print	Print Earth Engine objects.

—————————	—————————————————————————————————

XVIII. Utils functions

—————————	—————————————————————————————————

ee_utils_py_to_r	Convert between Python and R objects.

ee_utils_pyfunc	Wrap an R function in a Python function with the same signature.

ee_utils_shp_to_zip	Create a zip file from an sf object.

ee_utils_create_json	Convert a R list into a JSON file.

ee_utils_create_manifest_image	Create a manifest to upload an image.

ee_utils_create_manifest_table	Create a manifest to upload a table.

ee_utils_get_crs	Convert EPSG, ESRI or SR-ORG code into a OGC WKT.

ee_utils_future_value	The value of a future or the values of all elements in a container.

ee_utils_dataset_display	Search into the Earth Engine Data Catalog.

—————————	—————————————————————————————————

Acknowledgments

We want to offer a special thanks to Justin Braaten for his wise and helpful comments in the whole development of rgee. As well, we would like to mention the following third-party R/Python packages for contributing indirectly to the improvement of rgee:

gee_asset_manager - Lukasz Tracewski

geeup - Samapriya Roy

geeadd - Samapriya Roy

eemont - David Montero Loaiza

cartoee - Kel Markert

geetools - Rodrigo E. Principe

landsat-extract-gee - Loïc Dutrieux

earthEngineGrabR - JesJehle

sf - Edzer Pebesma

stars - Edzer Pebesma

gdalcubes - Marius Appel

See also

Author

Maintainer: Cesar Aybar csaybar@gmail.com (ORCID)

Other contributors:

Wu Qiusheng giswqs@gmail.com (ORCID) [contributor]

Lesly Bautista leslyarcelly.213@gmail.com (ORCID) [contributor]

Roy Yali ryali93@gmail.com (ORCID) [contributor]

Antony Barja antony.barja8@gmail.com (ORCID) [contributor]

Kevin Ushey kevin@rstudio.com [contributor]

Jeroen Ooms jeroen@berkeley.edu (ORCID) [contributor]

Tim Appelhans tim.appelhans@gmail.com [contributor]

JJ Allaire jj@rstudio.com [contributor]

Yuan Tang terrytangyuan@gmail.com [contributor]

Samapriya Roy samapriya.roy@gmail.com [contributor]

MariaElena Adauto 2a.mariaelena@gmail.com (ORCID) [contributor]

Gabriel Carrasco gabriel.carrasco@upch.pe (ORCID) [contributor]

Henrik Bengtsson henrikb@braju.com [contributor]

Jeffrey Hollister hollister.jeff@epa.gov (Hollister reviewed the package for JOSS, see https://github.com/openjournals/joss-reviews/issues/2272/) [reviewer]

Gennadii Donchyts (Gena reviewed the package for JOSS, see https://github.com/openjournals/joss-reviews/issues/2272/) [reviewer]

Marius Appel marius.appel@uni-muenster.de (Appel reviewed the package for JOSS, see https://github.com/openjournals/joss-reviews/issues/2272/) [reviewer]