Convert an sf object to EE object

Convert an sf object to EE object

sf_as_ee(x, check_ring_dir)

# S3 method for character
sf_as_ee(x, check_ring_dir = FALSE, ...)

# S3 method for sf
sf_as_ee(x, check_ring_dir = FALSE)

# S3 method for sfc
sf_as_ee(x, check_ring_dir = FALSE)

# S3 method for sfg
sf_as_ee(x, check_ring_dir = FALSE)

Arguments

x
sf object to be converted into a EE object.

check_ring_dir
logical. See st_read and details.

...
st_read arguments might be included.

Details

The conversion from sf to EE is a two-step process. First, sf_as_ee transform sf objects into a GeoJSON format using geojson_json. Second, the GeoJSON generated will be encrusted in an HTTP request using the server-side objects (ee$Geometry$*). If the sf object is a large spatial object (>1Mb) it is likely to cause bottlenecks and plodding connections. See Client vs Server documentation for more details. For dealing with very large spatial objects, it is recommended to import it into the GEE asset. See ee_upload for creating uploaded pipelines.

Earth Engine is strict on polygon ring directions (outer ring counter-clockwise, and the inner one clockwise). If check_ring_dir is TRUE, it check every ring, and revert them if necessary, to counter clockwise for outer, and clockwise for inner (hole) ones. By default this is FALSE because it is an expensive operation.

Examples

library(rgee)
library(sf)

ee_reattach() # reattach ee as a reserved word
ee_Initialize()

#> ── rgee 0.2.0 ──────────────────────── earthengine-api 0.1.213 ──

#>  ✔ email: not_defined
#>  ✔ Initializing Google Earth Engine: ✔ Initializing Google Earth Engine:  DONE!
#>  ✔ Earth Engine user: users/datacolecfbf 

#> ─────────────────────────────────────────────────────────────────

# sf
x <- st_read(system.file("shape/nc.shp", package = "sf")) %>%
  st_transform(4326)

#> Reading layer `nc' from data source `/home/aybarpc01/R/x86_64-pc-linux-gnu-library/3.6/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> epsg (SRID):    4267
#> proj4string:    +proj=longlat +datum=NAD27 +no_defs

ee_x <- sf_as_ee(x, check_ring_dir = TRUE)
ee_map(ee_x)

#>  vizparams is missing, the following params are assigned. 
#>  vizparams: {color:'000000', strokeWidth:3, pointRadius:3}
#> center is missing, the centroid of this EE FeatureCollection is used: 
#>  center:  -79.4004113801615 35.5595134054197